% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_script.R
\name{adapt_script}
\alias{adapt_script}
\title{adapt old scripts to package bundeswaldinventur}
\usage{
adapt_script(file_names = NA, path = ".",
  pattern = ".*\\\\.[RrSs]$|.*\\\\.[RrSs]nw$", all.files = TRUE,
  recursive = TRUE, ignore.case = FALSE, overwrite = FALSE,
  backup = overwrite, verbose = TRUE, clean_warning = FALSE)
}
\arguments{
\item{file_names}{character vector naming the scripts to be adapted.}

\item{path}{see \code{\link{find_files}}.}

\item{pattern}{see \code{\link{find_files}}.}

\item{all.files}{see \code{\link{find_files}}.}

\item{recursive}{see \code{\link{find_files}}.}

\item{ignore.case}{see \code{\link{find_files}}.}

\item{overwrite}{overwrite files on disk?}

\item{backup}{backup files to be overwriten?}

\item{verbose}{be verbose?}

\item{clean_warning}{delete lines containing only a "options(warn=2)"
instruction?}
}
\value{
invisibly the content of the change files.
}
\description{
Old scripts were based on a set of R codes in a customization directory that
were controlled by customization/customization.R. These are adapted to use
the package bundeswaldinventur.
}
