% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{add_colSums_prettify_and_print_xtable}
\alias{add_colSums_prettify_and_print_xtable}
\title{print an enhanced xtable from a data.frame.}
\usage{
add_colSums_prettify_and_print_xtable(data_frame, caption = "XXX")
}
\arguments{
\item{data_frame}{The data which to xtable::xtable().}

\item{caption}{A caption for the LaTeX table.}
}
\value{
The latex code created.
}
\description{
'Enhanced' means: add colSums(), use prettyNum()s.
}
\note{
The LaTeX table will have a latex label tab:data_frame. So you really
should avoid data_frame specifications that are not valid LaTeX names, like
in the examples.
}
\examples{
add_colSums_prettify_and_print_xtable(utils::head(airquality * 10^4))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
