% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits_and_pieces.R
\name{add_colSums_to_data_frame}
\alias{add_colSums_to_data_frame}
\title{add colSums to a data.frame, preserve its row.names}
\usage{
add_colSums_to_data_frame(data_frame, german = TRUE)
}
\arguments{
\item{data_frame}{The data from to add colSums to.}

\item{german}{If TRUE, the name of the colSums row will be 'Summe' and 'Sum'
otherwise.}
}
\value{
the modified data.frame.
}
\description{
add colSums to a data.frame, preserve its row.names
}
\examples{
add_colSums_to_data_frame(utils::head(airquality), german = FALSE)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
