% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{akl.lab.fun}
\alias{akl.lab.fun}
\title{Erzeugt Altersklassen-Beschriftung}
\usage{
akl.lab.fun(A.klass, A.k)
}
\arguments{
\item{A.klass}{vorgegebene Klassifizierungsregel.}

\item{A.k}{Anzahl der Altersklassen.}
}
\value{
Label fuer Altersklassen-Beschriftung.
}
\description{
Erzeugt Label fuer die Altersklassen-Beschriftung aus der
Klassififizierungsregel und der Anzahl an Altersklassen.
}
\section{Erstellungsdatum}{
 16.07.2014
}

\section{Note}{
 Anzahl der Altersklassen wird vom rufenden Programm berechnet.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
