% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{all_data_by_group}
\alias{all_data_by_group}
\title{Predicts value of a variable grouped by a tree species, BWI and owner}
\usage{
all_data_by_group(abbreviation, data_prefix = "FVBN.bagr")
}
\arguments{
\item{abbreviation}{String which defines the variable which shall be
predicted.}

\item{data_prefix}{The prefix from which the lists names are to be
established.}
}
\value{
Data frame with prediction for the given \code{abbreviation}, grouped
 by tree species group, BWI and owner.
}
\description{
Function predicts a variable and it's standard error grouped by tree species
and BWI.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
