% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits_and_pieces.R
\name{as_percent_and_round}
\alias{as_percent_and_round}
\title{round() numbers and multiply them with 100}
\usage{
as_percent_and_round(x, digits = 1)
}
\arguments{
\item{x}{the numbers to be modified.}

\item{digits}{The digits passed to round().}
}
\value{
the modified numbers.
}
\description{
The majority of readers can't cope with 0.1212. So we
change it to 12.1.
}
\examples{
as_percent_and_round(airquality[1:10, "Ozone"] / max(airquality[["Ozone"]],
                     na.rm = TRUE))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
