% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{ascii_umlauts}
\alias{ascii_umlauts}
\title{ASCII code for german umlauts}
\usage{
ascii_umlauts(string)
}
\arguments{
\item{string}{Non ASCII code for german umlaut.}
}
\value{
ASCII code for german umlaut.
}
\description{
Gets ASCII code for german umlauts.
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
