% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{ba.klass.fun}
\alias{ba.klass.fun}
\title{Prueft ob Baumart in Baumartengruppen vorkommt}
\usage{
ba.klass.fun(ba, ba.klass)
}
\arguments{
\item{ba}{BWI-Baumart-Code.}

\item{ba.klass}{Liste mit Baumartenzusammenfassung zu Baumgruppen \cr
Beispiel ba.klass <- list(c(10:19,30:39,90:99), c(20:29,40,50,51), c(100),
c(110,111), c(112,120:199),c(200:299)).}
}
\value{
Wenn der Baumartencode in der Liste mit der Baumartenzusammenfassung
 gefunden wurde, wird die Position des ersten Vectors in der Liste, welcher
 die Baumart enthaelt, zurueck gegeben. Wenn der Baumartencode nicht gefunden
 wurde, wird NA zurueck gegeben.
}
\description{
Funktion prueft ob ein bestimmter BWI-Baumarten-Code in einer Liste mit
Baumartenzusammenfassung zu Baumgruppen enthalten ist.
}
\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
