% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{ba.klass.lab.tab.fun}
\alias{ba.klass.lab.tab.fun}
\title{Erzeugt Zuordnungstabelle mit Baumartengruppen-Bezeichner}
\usage{
ba.klass.lab.tab.fun(ba.lab.klass, bacodes = get_data("bacode"))
}
\arguments{
\item{ba.lab.klass}{Liste mit Angaben zu Baumartenlabel und Baumartencode \cr
Beispiel: ba.lab.klass <- list(ba.lab = c("FiTa","DglKiLae","Bu","Ei","BLb",
"WLb"), ba.code = list(c(10:19,30:39,90:99), c(20:29,40,50,51), c(100),
c(110,111), c(112:199), c(200:299))).}

\item{bacodes}{Basistabelle mit Baumartencodes.}
}
\value{
Dataframe-Tabelle mit Baumartengruppen-Nummer und dem entspechenden
 Baumartengruppen-Label.
}
\description{
Funktion erzeugt Zuordnungstabelle mit Baumartengruppen-Bezeichner, welche
benoetigt wird, um Klassifikation nach frei zu definierenden Baumartengruppen
zu ermoeglichen.
}
\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
