% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{biotop.baeume.fun}
\alias{biotop.baeume.fun}
\title{Funktion zur Auswertung der Biotop-Baeume}
\usage{
biotop.baeume.fun(wzp4.merkmale, baeume, ecken, trakte, A, auswahl)
}
\arguments{
\item{wzp4.merkmale}{Tabelle, welche die besonderen Biotop-Baum-Merkmale
enthaelt u.a. FaulKon, Hoehle, Bizarr, Uralt, Horst, MBiotop.}

\item{baeume}{Probebaum-Attribute.}

\item{ecken}{Trakteckenmerkmale.}

\item{trakte}{traktbezogene Informationen (u.a. m, m_HB, ...).}

\item{A}{Flaeche des gesamten Inventurgebietes in ha.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und genutzten Baeumen(?)), \strong{Stratum}
 (\code{auswahl}), \strong{nTE} (Anzahl Ecken im Stratum), \strong{HBF}
 (Holzbodenflaeche in ha), \strong{se.HBF} (Standardfehler HBF), \strong{BL}
 (Flaeche der Bloessen in ha), \strong{se.BL} (Standardfehler BL),
 \strong{iBL} (Flaeche der idellen Bloessen ("Luecken") in ha),
 \strong{se.iBL} (Standardfehler iBL), \strong{LK} (relative
 Lueckenkorrektur), \strong{se.LK}  (Standardfehler LK),
 \strong{Biotopbaumkennwerte} (Dataframetabelle mit Spalten fuer Attribute,
 Wert und Standardfehler. Folgende Attribute werden aggregiert:
 Gesamtstueckzahl, Stueckzahl je ha, gesamtes Volumen, Volumen je ha,
 Stueckvolumen, gesamtes Volumen der Laubbaeume, Laubbaumvolumen je ha,
 gesamtes Nadelbaumvolumen, Nadelbaumvolumen je ha, gesamte oberirdische
 Biomasse, oberirdische Biomasse je ha).
}
\description{
Funktion wertet Biotop-Baeume aus, die bei der 3. BWI als Merkmal der
Probebaeume in der WZP 4 aufgenommen worden sind. Als Biotop-Baeume werden
die Probebaeume erfasst, welche mindestens eines der folgenden Merkmale
erfuellen: FaulKon, Hoehle, Bizarr, Uralt, Horst, MBiotop.
}
\section{Erstellungsdatum}{
 10.10.2014
}

\section{Hinweis}{
 funktioniert nicht fuer <baume.1> der BWI 1!!!
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
