% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bwi_data}
\alias{bwi_data}
\alias{bacode}
\alias{baeume.1}
\alias{baeume.2}
\alias{baeume.3}
\alias{baeume.12}
\alias{baeume.23}
\alias{ecke.feld.3}
\alias{ecken.1}
\alias{ecken.2}
\alias{ecken.3}
\alias{ecken.23}
\alias{fba.2}
\alias{fba.3}
\alias{gk_coord}
\alias{kreise}
\alias{ntns.te}
\alias{schutzgebiete.3}
\alias{totholz.2}
\alias{totholz.3}
\alias{trakte.1}
\alias{trakte.2}
\alias{trakte.3}
\alias{verj.2}
\alias{verj.3}
\alias{verj.kl4m.2}
\alias{verj.kl4m.3}
\alias{wzp4.merkmale.2}
\alias{wzp4.merkmale.3}
\alias{x_dichte}
\alias{x_fba.3}
\title{BWI Data}
\format{A data frame with 88 rows and 17 variables:
\describe{
  \item{ICode}{fix me}
  \item{ACode}{fix me}
  \item{KurzD}{fix me}
  \item{Konv_Ba2_Ba3}{fix me}
  \item{AutoKonv_Ba2_Ba3}{fix me}
  \item{BaGr}{abbreviation of the tree species group}
  \item{BDAT.BA}{abbreviation of the specific tree species}
  \item{BDAT.BANr}{number of the specific tree species}
  \item{EHK.BA}{fix me}
  \item{StFl.BA}{fix me}
  \item{BA_STR}{fix me}
  \item{Krenn_HT.BA}{fix me}
  \item{WKoeff_BA}{fix me}
  \item{WKoeff_BA_Code}{fix me}
  \item{Kommentar}{comment}
  \item{WKoeff_BA_Code_alt}{fix me}
  \item{WKoeff_BA_BWI3}{fix me}
}}
\usage{
bacode

baeume.1

baeume.2

baeume.3

baeume.12

baeume.23

ecke.feld.3

ecken.1

ecken.2

ecken.3

ecken.23

fba.2

fba.3

gk_coord

kreise

ntns.te

schutzgebiete.3

totholz.2

totholz.3

trakte.1

trakte.2

trakte.3

verj.2

verj.3

verj.kl4m.2

verj.kl4m.3

wzp4.merkmale.2

wzp4.merkmale.3

x_dichte

x_fba.3
}
\description{
All data resampled from \pkg{bwibw}.

Image of an SQL table giving the tree species codes.

Data frame with tree information from BWI1.

Data frame with tree information from BWI2.

Data frame with tree information from BWI3.

Data frame with tree information from BWI1 and BWI2.

Data frame with tree information from BWI2 and BWI3.

Tabelle enthaelt Informationen ueber die Waldlebensraumtypen, aufgenommen in
der BWI 3.

Eckenbezogene Informationen aus der Traktvorklaerung der BWI 1.

Eckenbezogene Informationen aus der Traktvorklaerung der BWI 2.

Eckenbezogene Informationen der Traktvorklaerung der BWI 3.

Kombinierte Eckeninformationen aus den Traktvorklaerungen der BWI 2 and BWI
3.

Forstlich besonders bedeutsame Pflanzenarten (FBA) der BWI 2 je Traktecke.

Forstlich besonders bedeutsame Pflanzenarten (FBA) der BWI 3 je Traktecke.

Tabelle enthaelt Informationen zu Sollkoordinaten und tatsaechlich
eingemessenen Koordinaten.

Tabele enthaelt Infomrationen ueber Kreise, die entsprechenden Kreiscodes der
BWI, die Kreishauptstadt und die Groesse des Kreises.

Angaben zur Naturnaehestufe pro Traktecke.

Informationen ueber die Schutzgebiete zum Zeitpunkt der BWI 3.

Tabelle mit Information zum Totholz aufgebnommen waehrend der BWI 2.

Tabelle mit Information zum Totholz aufgebnommen waehrend der BWI 3.

Traktbezogene Informationen zu den Trakten der BWI 1.

Traktbezogene Informationen zu den Trakten der BWI 2.

Traktbezogene Informationen zu den Trakten der BWI 3.

Informationen zur Verjuengung der BWI 2.

Informationen zur Verjuengung der BWI 3.

Informationen zur Verjuengung, welche kleiner als 4 Meter ist, der BWI 2.

Informationen zur Verjuengung, welche kleiner als 4 Meter ist, der BWI 3.

Informationen zu Schaeden, die bei der Winkelzaehlprobe der BWI 2 aufgenommen
wurden.

Informationen zu Schaeden, die bei der Winkelzaehlprobe der BWI 3 aufgenommen
wurden.

Uebersichtstabelle fuer Dichteangaben

Uebersichtstabelle zu den Forstlich besonders bedeutsame Pflanzenarten
}
\section{moegliche FBA}{

 11 - 	Adlerfarn \cr
 12 -	Brennessel \cr
 13 -	Riedgras \cr
 14 -	Honiggras \cr
 15 -	Reitgras \cr
 16 - 	Heidekraut \cr
 17 - 	Heidelbeere \cr
 18 -	Brombeere \cr
 19 - 	Riesenbaerenklau \cr
 20 -	Riesenknoeterich \cr
 21 -	Druesiges Springkraut \cr
 22 -	Kleinbluetiges Springkraut \cr
 23 -	Kernesbeere


 11 -   Adlerfarn \cr
 12 -	Brennessel \cr
 13 -	Riedgras \cr
 14 -	Honiggras \cr
 15 -	Reitgras \cr
 16 - 	Heidekraut \cr
 17 - 	Heidelbeere \cr
 18 -	Brombeere \cr
 19 - 	Riesenbaerenklau \cr
 20 -	Riesenknoeterich \cr
 21 -	Druesiges Springkraut \cr
 22 -	Kleinbluetiges Springkraut \cr
 23 -	Kernesbeere
}

\keyword{datasets}
