% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_species_shares.R
\name{cast_tree_species_percent}
\alias{cast_tree_species_percent}
\title{Calculates relative tree species area}
\usage{
cast_tree_species_percent(statistics_list)
}
\arguments{
\item{statistics_list}{String which defines the list holding the statistical
data.}
}
\value{
Data Frame with information about relative tree species area for each
 species. Information for all three BWI's.
}
\description{
Function calculates the relative tree species area for each tree species
group for all three BWI's. BWI's just given through the different years:
1987, 2002 and 2012.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
