% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadwood.R
\name{deadwood}
\alias{deadwood}
\title{Aggregate deadwood amount}
\usage{
deadwood(totholz, v = c("tnr", "enr"), bwi = 3)
}
\arguments{
\item{totholz}{Dataframe table with information about deadwood.}

\item{v}{Vector with Columns on which analysis shall be done. By default set
to c("tnr", "enr").}

\item{bwi}{Number of BWI. By default set to 3.}
}
\value{
Dataframe- table with input columns and an added column with contains
 information about deadwood amount.
}
\description{
Function aggregates amount of deadwood by following calculation:
\code{totholz$tvol * totholz$thf * totholz$anz}
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
