% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{dkl.lab.fun}
\alias{dkl.lab.fun}
\title{Erzeugt Durchmesserklassen-Beschriftung}
\usage{
dkl.lab.fun(D.klass, D.k)
}
\arguments{
\item{D.klass}{vorgegebene Klassifizierungsregel.}

\item{D.k}{Anzahl der Durchmesserklassen.}
}
\value{
Label fuer Durchmesserklassen-Beschriftung.
}
\description{
Erzeugt ein Label fuer die Durchmesserklassen-Beschriftung aus der
Klassifizierungsregel und der Anzahl der Durchmesserklassen.
}
\section{Erstellungsdatum}{
 16.07.2014
}

\section{Korrekturdatum}{
 07.02.2015
}

\section{Note}{
 Anzahl der Durchmesserklassen wird vom rufenden Programm
 errechnet.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
