% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadwood.R
\name{drop_variables}
\alias{drop_variables}
\title{Selects Columns from deadwood table.}
\usage{
drop_variables(totholz)
}
\arguments{
\item{totholz}{Dataframe-table with information about deadwood.}
}
\value{
Subset of input table.
}
\description{
Function selects following columns from deadwood table: tnr, tvol, thf, anz,
tart, tbagr, tzg. All other columns are droped.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
