% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{fba.stratum.fun}
\alias{fba.stratum.fun}
\title{Berechnet Verteilung der Dichtestufen}
\usage{
fba.stratum.fun(fba, fba.tab, auswahl, ecken, trakte, A)
}
\arguments{
\item{fba}{Vektor mit den auszuwertenden FBA, wird NA oder "alle" uebergeben,
werden alle ausgewertet.}

\item{fba.tab}{Die bei der BWI 3 erfolgte FBA-Dichte-Ansprachen je Traktecke.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt.}

\item{ecken}{Ecken-Tabelle.}

\item{trakte}{Trakte-Tabelle.}

\item{A}{Flaeche des Inventurgebietes in ha.}
}
\value{
Liste mit folgenden Attributen: \strong{Stratum}, \strong{HBF}
 (Holzbodenflaeche), \strong{se.HBF} (Standardfehler Holzbodenflaeche),
 \strong{FBA}, \strong{Dichtestufen}, \strong{FBA.Dichtevertlg},
 \strong{FBA.Flaeche}. Wobei FBA.Dichtevertlg ein Array mit den
 Dichteanteilen der FBA ist und FBA.Flaeche ein Array ist, welches Wert und
 Standardfehler fuer Flaeche (ha) und Flaechenanteile enthaelt.
}
\description{
Funktion berechnet Verteilung der Dichtestufen fuer Forstlich bedeutsame
Arten (FBA) entsprechend der Definition der BWI 3; die Funktion kann auch
fuer die entsprechende FBA-Tabelle der BWI 2 <fba.2> angewandt werden
(dort sind aber nur die Arten 11 bis 18 erfasst).
}
\section{Version}{
 01.05.2015
}

\section{Note}{
 Folgende FBA sind bei BWI 3 definiert:Adlerfarn, Brennessel,
 Riedgras, Honiggras, Reitgras, Heidekraut, Heidelbeere, Brombeere,
 Riesenbaerenklau, Riesenknoeterich, Druesiges Springkraut, Kleinbluetiges
 Springkraut, Kermesbeere; mit Code 11:23.
}

\section{Hinweis}{
 fehlende Angaben ("keine Angabe") werden als Nicht-
 Vorkommen interpretiert. \cr
 Zur Flaechenberechnung: Die Beobachtung des Vorkommens wird in visuell
 angesprochene Dichtestufen angegeben: 0: "nicht vorhanden", 1: "selten, bis
 10%", 2: "haeufig, > 10 bis 50%", 3:"flaechig, > 50%". Diese Dichten werden in
 Flaechenanteile umgerechnet: 0: 0; 1: 0.05; 2: 0.3; 3: 0.7; da jede
 Stichprobe mit dem Repraesentationsfaktor (RF) hochgerechnet wird, ergeben
 sich hieraus Flaechenschaetzungen in ha
 d.fl.ant <- data.frame(dichte=0:3,fl.ant=c(0,0.05,0.3,0.7))
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
