% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{fl.proz.stratum.fun}
\alias{fl.proz.stratum.fun}
\title{Berechnet den relativen Anteil einer Untermenge}
\usage{
fl.proz.stratum.fun(stratum, substratum, ecken,
  trakte.3 = get_data("trakte.3"))
}
\arguments{
\item{stratum}{Liste, welche die Eckenmerkmale enthaelt nachdem das
uebergeordnete Stratum (Obermenge) definiert wird.}

\item{substratum}{Liste, welche die Eckenmermale enthaelt nachdem das
Substratum (Untermenge) definiert wird.}

\item{ecken}{Traktecken-Merkmale, die zur Stratenbildung dienen.}

\item{trakte.3}{Trakt information.}
}
\value{
Liste mit relativen Anteil des Substratums am Stratum und dessen
 Standardfehler.
}
\description{
Funktion berechnet den relativen Anteil des \code{substratums} im
\code{stratum}, welches das uebergeordnete Stratum (Obermenge) definiert.
}
\section{version}{
 18.11.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
