% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl.stratum.fun.R
\name{fl.stratum.fun}
\alias{fl.stratum.fun}
\title{Berechnet die Flaeche entsprechend einer Auswahl}
\usage{
fl.stratum.fun(auswahl, ecken, trakte, A, trakte.3 = get_data("trakte.3"))
}
\arguments{
\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt.}

\item{ecken}{Traktecken-Merkmale.}

\item{trakte}{Trakt-Kennwerte (m, m_HB, m_bHB, m_Wa).}

\item{A}{Flaeche des Inventurgebietes in ha.}

\item{trakte.3}{Trakt information.}
}
\value{
Liste mit der Flaeche und dem Standardfehler.
}
\description{
Funktion berechnet die Flaeche entsprechend der Definition in \code{auswahl}.
}
\section{Version}{
 30.07.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
