% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR_Landkreis.R
\name{flaechen.kreis.fun.1}
\alias{flaechen.kreis.fun.1}
\title{Aggregiert Flaechen fuer den Landkreis}
\usage{
flaechen.kreis.fun.1(kreiscode, A = get_design("a", 3),
  A.12 = get_design("a", 1), kreise = get_data("kreise"),
  trakte.3 = get_data("trakte.3"), ecken.3 = get_data("ecken.3"),
  trakte.2 = get_data("trakte.2"), ecken.2 = get_data("ecken.2"),
  trakte.1 = get_data("trakte.1"), ecken.1 = get_data("ecken.1"))
}
\arguments{
\item{kreiscode}{Code fuer den Landkreis.   kreiscode}

\item{A}{Area for BWI 3.}

\item{A.12}{Area for BWI1 and BWI2}

\item{kreise}{County information.}

\item{trakte.3}{Trakt information.}

\item{ecken.3}{Ecken information.}

\item{trakte.2}{Trakt information.}

\item{ecken.2}{Ecken information.}

\item{trakte.1}{Trakt information.}

\item{ecken.1}{Ecken information.}
}
\value{
Liste mit folgenden Komponenten: \strong{Kreis} (Label fuer den
 Kreis), \strong{KreisCode} (\code{kreiscode}),
 \strong{Kreisflaeche_gesamt_ha} (Groesse des Kreises in ha),
 \strong{Waldanteil} (waldanteil des Kreises an den verschiedenen BWIs),
 \strong{Waldflaeche} (Waldflaeche und Standardfehler fuer verschiedene
 Eigentumsklassen und BWIs), \strong{Waldfl_Proz} (Waldflaeche in Prozent und
 Standardfehler fuer verschiedene Eigentumsklassen), \strong{begehb_Holzbfl}
 (Holzbodenflaeche und Standardfehler fuer verschiedene Eigentumsklassen)
}
\description{
Funktion wertet verschiedene Flaechen fuer den Landkreis aus, unterteilt fuer
die Eigentumskategorien.
}
\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
