% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR_Landkreis.R
\name{flaechen.stratum.fun.1}
\alias{flaechen.stratum.fun.1}
\title{Aggregiert Flaechen fuer ein Stratum}
\usage{
flaechen.stratum.fun.1(auswahl, A = get_design("a", 3),
  A.12 = get_design("a", 1), trakte.3 = get_data("trakte.3"),
  ecken.3 = get_data("ecken.3"), trakte.2 = get_data("trakte.2"),
  ecken.2 = get_data("ecken.2"), trakte.1 = get_data("trakte.1"),
  ecken.1 = get_data("ecken.1"))
}
\arguments{
\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.:
list(Wa=c(3,5), Begehbar=1).}

\item{A}{Area for BWI 3.}

\item{A.12}{Area for BWI1 and BWI2}

\item{trakte.3}{Trakt information.}

\item{ecken.3}{Ecken information.}

\item{trakte.2}{Trakt information.}

\item{ecken.2}{Ecken information.}

\item{trakte.1}{Trakt information.}

\item{ecken.1}{Ecken information.}
}
\value{
Liste mit folgenden Komponenten: \strong{Stratum} (\code{auswahl}),
 \strong{Waldflaeche} (Waldflaeche und Standardfehler fuer verschiedene
 Eigentumsklassen und BWIs), \strong{Waldfl_Proz} (Waldflaeche in Prozent und
 Standardfehler fuer verschiedene Eigentumsklassen), \strong{begehb_Holzbfl}
 (Holzbodenflaeche und Standardfehler fuer verschiedene Eigentumsklassen)
}
\description{
Funktion berechnet verschiedene Flaechen nach Eigentumskategorien.
}
\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
