% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvbn2a.R, R/fvbn2b.R, R/fvbn2c.R, R/fvbn2d.R,
%   R/fvbn2e.R
\name{FVBN.bagrupp.akl.dkl.stratum.fun.2a}
\alias{FVBN.bagrupp.akl.dkl.stratum.fun.2a}
\alias{FVBN.bagrupp.akl.dkl.stratum.fun.2b}
\alias{FVBN.bagrupp.akl.dkl.stratum.fun.2c}
\alias{FVBN.bagrupp.akl.dkl.stratum.fun.2d}
\alias{FVBN.bagrupp.akl.dkl.stratum.fun.2e}
\title{Zustandshochrechnung}
\usage{
FVBN.bagrupp.akl.dkl.stratum.fun.2a(baeume, ecken, trakte, A, inv, BA.grupp,
  A.klass, D.klass, auswahl)

FVBN.bagrupp.akl.dkl.stratum.fun.2b(baeume, ecken, trakte, A, inv, BA.grupp,
  A.klass, D.klass, auswahl)

FVBN.bagrupp.akl.dkl.stratum.fun.2c(baeume, ecken, trakte, A, inv, BA.grupp,
  A.klass, D.klass, auswahl)

FVBN.bagrupp.akl.dkl.stratum.fun.2d(baeume, ecken, trakte, A, inv, BA.grupp,
  A.klass, D.klass, auswahl)

FVBN.bagrupp.akl.dkl.stratum.fun.2e(baeume, ecken, trakte, A, inv, BA.grupp,
  A.klass, D.klass, auswahl)
}
\arguments{
\item{baeume}{Datentabelle mit Bauminformationen. Die Tabelle muss mindestens
die Attribute TNr, ENr, BA, Alt., BHD., VolV., oiB., NHa., StFl. enthalten
(. steht fuer 1 oder 2, je nachdem welcher Zustand ausgewertet werden soll),
man kann auch die selektierten Attribute mit Namen ohne Kennziffer
uebergeben, wenn bereits eine eindeutige Auswahl der Attribute in
\code{baeume} uebergeben wird.}

\item{ecken}{Eckenmerkmale.}

\item{trakte}{Traktmerkmale.}

\item{A}{Gesamtflaeche in ha des Inventurgebiets zum jeweiligen
Inventurzeitpunkt (sollte eigentlich konstant sein).}

\item{inv}{Inventur 1:= Vorinventur, 2:= Folgeinventur, wenn eine
\code{baeume}-Tablle mit Attributen aus zwei Aufnahmezeitpunkten uebergeben
wird. Wenn dies der Fall ist, dann entsprechen die jeweils fuer die
auszuwertende Inventur relevanten Attribute der Konvention: Aktuelle
("Folge") Inventur hat im Attributnamen eine 2 am Ende, also "Alt2", "BHD2",
"D032", "H2", "HSt2", "VolV2", "VolE2", "oiB2", "StFl2", "NHa2", die
Vorinventur eine 1; d.h. im Falle der BWI 3 haben die Attributbezeichner die
2 im Namen, die Daten der Vorinventur BWI 2 die 1. Um mit denselben
Algorithmen fuer beide Zustaende arbeiten zu koennen, werden fuer den
jeweiligen Auswertungsfall \code{inv} (Vorinventur = 1 bzw. Folge-(aktuelle)
Inventur =2 die Attibutnamen "neutralisiert", also die Kennung 1 oder 2 im
Attributnamen entfernt.}

\item{BA.grupp}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199),c(200:299))).}

\item{A.klass}{Liste mit den Klassifizierungsparametern fuers Alter: z.B.
list(A.ob=160, A.b=20).}

\item{D.klass}{Liste mit den Klassifizierungsparametern fuer Durchmesser z.B.
list(D.unt=0, D.ob=70, D.b=10, Ndh=T), Ndh (Nicht-Derbholz) = T bedeutet,
dass zusaetzlich Nicht-Dh (unter 7 cm) ausgewiesen wird, sonst gilt
\code{D.unt} als unterste Schwelle.}

\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.:
list(Wa=c(3,5), Begehbar=1).}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und Version.baeume.b, \strong{Stratum}
 (\code{auswahl}), \strong{nTE} (Anzahl Ecken im Stratum), \strong{HBF}
 (Holzbodenflaeche in ha), \strong{se.HBF} (Standardfehler HBF), \strong{BL}
 (Flaeche der Bloessen in ha), \strong{se.BL} (Standardfehler BL),
 \strong{iBL} (Flaeche der idellen Bloessen ("Luecken") in ha),
 \strong{se.iBL} (Standardfehler iBL), \strong{LK} (relative
 Lueckenkorrektur), \strong{se.LK}  (Standardfehler LK), \strong{Attribute1}
 (Vektor mit berechneten absoluten Attributnamen), \strong{Attribute2}
 (vektor mit berechneten pro ha Attributname), \strong{Groessen} (Vektor mit
 berechneten Groessen), \strong{BAGR} (Labels fuer Baumartengruppen aus
 \code{ba.grupp}), \strong{AKL} (Labels der Altersklassen), \strong{DKL}
 (Labels der Durchmesserklassen), \strong{T.FVBN.Bagr.Akl.Dkl} (Array
 mit berechneten Groessen (Wert und Standardfehler) fuer Gesamtwerte jeweils
 fuer 9 Baumartengruppen, Alters- und Durchmesserklassen),
 \strong{FVBN.ha.Bagr.Akl.Dkl} (Array mit berechneten Groessen (Wert und
 Standardfehler) fuer hektabezogene Kenngroessen), \strong{nT.bagr.Akl.Dkl}
 (Anzahl Traktecken (?)).
}
\description{
Funktion wertet nach frei definierbaren BA-Gruppen, Alters- und Durchmesser-
Klassen im Stratum aus. Die berechneten Zielgroessen sind Flaeche, Vorrat,
Biomasse und Stammzahl (N). Variante mit Varianzschaetzung ueber
Fehlerfortpflanzung.
}
\section{Functions}{
\itemize{
\item \code{FVBN.bagrupp.akl.dkl.stratum.fun.2a}: (version 2a)

\item \code{FVBN.bagrupp.akl.dkl.stratum.fun.2b}: (version 2b)

\item \code{FVBN.bagrupp.akl.dkl.stratum.fun.2c}: (version 2c)

\item \code{FVBN.bagrupp.akl.dkl.stratum.fun.2d}: (version 2d)
}}

\section{Erstellungsdatum}{
 13.03.2014
}

\section{Aktualisierungen}{

 13.03.2015 Korrektur \code{akl.lab} bzw. \code{dkl.lab} \cr
 22.02.2015 Erweiterung um "alle DKL" und "alle AKL" \cr
 02.12.2014 Erweiterung um alle Baumarten: wenn nach BAGR differenziert wird,
   wird zusatzlich die "Summen"- Auswertung fuer alle Baumarten aufgerufen
   und das Ergebnis in die Tabelle eingebaut \cr
 01.11.2014 Neue Version mit Stammzahlen im Hauptbestand \cr
 11.10.2014 Erweiterung um Baumartenanteile mit Fehler \cr
 19.04.2014 bei Ha-Bezug fuer Hauptbestands(HB)-Vorrat wird Luecken-Korrektur
   rueckgaengig gemacht, um Bezug nur zur HB-Flaeche zu erzielen
   (\code{T.bl}, \code{T.ibl}).
}

\section{Hinweis}{
 Verallgemeinerte Version fuer die Auswertung
 unterschiedlicher Inventurzeit-punkte, also BWI 1, BWI 2, BWI 3. \cr
 Die Funktion \code{\link{stratum.fun}} benoetigt desweiteren die Tabelle
 \code{bacode}. \cr
 Um Konflikte mit unterschiedlicher Gross- und Kleinschreibung bei den
 Attributnamen zu vermeiden, werden innerhalb dieser Funktion alle
 Attributnamen auf Kleinschreibung umgestellt.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
