% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{fvbn.kreis.fun.1}
\alias{fvbn.kreis.fun.1}
\title{Zustandshochrechnung fuer einen Landkreis}
\usage{
fvbn.kreis.fun.1(kreiscode, eig.list, bwi.list, bagr, A = get_design("a", 3),
  A.12 = get_design("a", 1), kreise = get_data("kreise"),
  trakte.3 = get_data("trakte.3"), ecken.3 = get_data("ecken.3"),
  baeume.3 = get_data("baeume.3"), trakte.2 = get_data("trakte.2"),
  ecken.2 = get_data("ecken.2"), baeume.2 = get_data("baeume.2"),
  trakte.1 = get_data("trakte.1"), ecken.1 = get_data("ecken.1"),
  baeume.1 = get_data("baeume.1"))
}
\arguments{
\item{kreiscode}{Code fuer den Landkreis.}

\item{eig.list}{Liste mit Eigentumskategorien (moegliche Kategorien: gw, stw,
kw, oew, pw, gpw, mpw, kpw).}

\item{bwi.list}{Liste mit BWI-Aufnahmen (1, 2, 3).}

\item{bagr}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199),c(200:299))).}

\item{A}{Area for BWI 3.}

\item{A.12}{Area for BWI1 and BWI2}

\item{kreise}{County information.}

\item{trakte.3}{Trakt information.}

\item{ecken.3}{Ecken information.}

\item{baeume.3}{Tree information.}

\item{trakte.2}{Trakt information.}

\item{ecken.2}{Ecken information.}

\item{baeume.2}{Tree information.}

\item{trakte.1}{Trakt information.}

\item{ecken.1}{Ecken information.}

\item{baeume.1}{Tree information.}
}
\value{
Dataframe-Tabelle mit FVBN-Auswertung nach Eigentumsklasse und
 aufgelisteter BWI.
}
\description{
Funktion berechnet Standard-FVBN-Auswertung fuer den Landkreis mit
\code{kreiscode} fuer die in der \code{eig.list} aufgefuehrten
Eigentumskategorien {gw,stw,kw,oew,pw,gpw,mpw,kpw} fuer die in
\code{bwi.list} aufgefuehrten BWI-Aufnahmen {1,2,3}.Es wird die Funktion
\code{\link{FVBN.bagrupp.akl.dkl.stratum.fun.2d}} verwendet, welche neben
Baumartengruppen auch die Summenwerte fuer alle Baumarten liefert.
}
\section{Erstellungsdatum}{
 02.12.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
