% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{fvbn.stratum.fun.2}
\alias{fvbn.stratum.fun.2}
\title{Zustandshochrechnung fuer ein Stratum}
\usage{
fvbn.stratum.fun.2(auswahl, eig.list, bwi.list, bagr, a.klass, d.klass,
  A = get_design("a", 3), A.12 = get_design("a", 1),
  trakte.3 = get_data("trakte.3"), ecken.3 = get_data("ecken.3"),
  baeume.3 = get_data("baeume.3"), trakte.2 = get_data("trakte.2"),
  ecken.2 = get_data("ecken.2"), baeume.2 = get_data("baeume.2"),
  trakte.1 = get_data("trakte.1"), ecken.1 = get_data("ecken.1"),
  baeume.1 = get_data("baeume.1"))
}
\arguments{
\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.:
list(Wa=c(3,5), Begehbar=1).}

\item{eig.list}{Liste mit Eigentumskategorien (moegliche Kategorien: gw, stw,
kw, oew, pw, gpw, mpw, kpw).}

\item{bwi.list}{Liste mit BWI-Aufnahmen (1, 2, 3).}

\item{bagr}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199),c(200:299))).}

\item{a.klass}{Liste mit den Klassifizierungsparametern fuers Alter: z.B.
list(A.ob=160, A.b=20).}

\item{d.klass}{Liste mit den Klassifizierungsparametern fuer Durchmesser z.B.
list(D.unt=0, D.ob=70, D.b=10, Ndh=T), Ndh (Nicht-Derbholz) = T bedeutet,
dass zusaetzlich Nicht-Dh (unter 7 cm) ausgewiesen wird, sonst gilt
\code{D.unt} als unterste Schwelle.}

\item{A}{Area for BWI 3.}

\item{A.12}{Area for BWI1 and BWI2}

\item{trakte.3}{Trakt information.}

\item{ecken.3}{Ecken information.}

\item{baeume.3}{Tree information.}

\item{trakte.2}{Trakt information.}

\item{ecken.2}{Ecken information.}

\item{baeume.2}{Tree information.}

\item{trakte.1}{Trakt information.}

\item{ecken.1}{Ecken information.}

\item{baeume.1}{Tree information.}
}
\value{
Dataframe-Tabelle mit FVBN-Auswertung nach Eigentumsklasse und
 aufgelisteter BWI.
}
\description{
Funktion berechnet Standard-FVBN-Auswertung fuer die Befundeinheit
\code{auswahl} fuer die in der \code{eig.list} aufgefuehrten
Eigentumskategorien {gw, stw, kw, oew, pw, gpw, mpw, kpw} sowie die in
\code{bwi.list} aufgefuehrten BWI-Aufnahmen {1,2,3}. Es wird die Funktion
\code{\link{FVBN.bagrupp.akl.dkl.stratum.fun.2d}} verwendet, welche neben
Baumartengruppen auch die Summenwerte fuer alle Baumarten liefert.
}
\section{Erstellungsdatum}{
 02.12.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
