% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{german_number}
\alias{german_number}
\title{Changes number formatting to german style}
\usage{
german_number(x, ...)
}
\arguments{
\item{x}{A number.}

\item{...}{arguemnts passed to \code{\link{format}}.}
}
\value{
number with changed formatting.
}
\description{
Changes number fformatting to german style - "." as big mark and "," as
decimal mark.
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
