% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_abbreviation_for_label}
\alias{get_abbreviation_for_label}
\title{Get the abbreviation of a label.}
\usage{
get_abbreviation_for_label(label)
}
\arguments{
\item{label}{The label which shall be replaced.}
}
\value{
The abbreviation as character. If \code{label} is not known, the
  function will return 'unknown label'.
}
\description{
Replace a BWI label with the appropriate abbreviation.
}
\examples{
label <- "Baumartenflaeche"
print(bundeswaldinventur::get_abbreviation_for_label(label))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
