% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_axis_label_for_abbreviation}
\alias{get_axis_label_for_abbreviation}
\title{Get axis label for abbreviation.}
\usage{
get_axis_label_for_abbreviation(abbreviation)
}
\arguments{
\item{abbreviation}{[character] A abbreviation which shall be replaced by an
axis label.}
}
\value{
An expression holding a label and an approriate unit or a label as
 charater. If \code{abbreviation} is not known, it will be returned.
}
\description{
Get the axis label for a abbreviation containing a label with the appropriate
unit.
}
\examples{
abbreviation <- "V_DhmR"
label <- get_axis_label_for_abbreviation(abbreviation)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
