% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_getters.R
\name{get_bwi_colors}
\alias{get_bwi_colors}
\title{get standard bwi plotting colors}
\usage{
get_bwi_colors()
}
\value{
a vector with rgb values
}
\description{
we use standard colors for BWI 1, 2, and 3,
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
