% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_getters.R
\name{get_bwi_species_groups}
\alias{get_bwi_species_groups}
\title{get bwi species group list}
\usage{
get_bwi_species_groups(type = "default")
}
\arguments{
\item{type}{Select different lists.}
}
\value{
a list with
\itemize{
    \item[\var{bagr.lab}] a vector of tree species group labels.
    \item[\var{ba.grupp}] a corresponding list of vectors giving the species
    codes for each tree species group.
    \item[\var{ba.colors}] a corresponding vector of color definitions.
    \item[\var{ba.text}] another vector of tree species group labels.
}
}
\description{
the list of species groups is used by many functions.
}
\section{warning}{
 there is a second function \code{\link{get_species_groups}}
which does nearly the same but the return value is slightly different.
}

\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
