% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_color_for_ownership}
\alias{get_color_for_ownership}
\title{Get color for ownership.}
\usage{
get_color_for_ownership(ownership)
}
\arguments{
\item{ownership}{The label describing the ownership.}
}
\value{
ownership_colors[index] Which holds the information about the color
  for the wanted ownership (\code{index}).
}
\description{
Get the color for a certain ownership to use in graphics.
}
\examples{
ownership <- "Mittlerer Privatwald"
color <- get_color_for_ownership(ownership)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
