% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_colors_for_ownership}
\alias{get_colors_for_ownership}
\title{Get colors for ownership.}
\usage{
get_colors_for_ownership(x)
}
\arguments{
\item{x}{A vector containing one or more ownerships.}
}
\value{
A vector containing the colors for the ownerships, indexed by the
  abbreviations for the ownerships.
}
\description{
Get the color for certain ownerships to use in graphics.
}
\examples{
ownership <- c("Kleinprivatwald", "Mittlerer Privatwald")
colors <- get_colors_for_ownership(ownership)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
