% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Load Data from \pkg{bwibw}}
\usage{
get_data(name)
}
\arguments{
\item{name}{a string of length one giving the name of the data.frame.}
}
\value{
the data requested.
}
\description{
Get data if we want to use example data provided with the package.
}
\details{
this is a wrapper reading get_options("fake_data") and returning an error if
this is not set to TRUE, hence globally en/dis-abling the reading of example
data into function arguments. Which I need because Gerald used global
data.frames in his functions.
}
\examples{
if (require("bwibw", quietly = TRUE)) get_data("bacode")
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
