% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_objects.R
\name{get_global_objects}
\alias{get_global_objects}
\title{Define Global Variables and Data Needed by Historical Analysis Scripts}
\usage{
get_global_objects()
}
\value{
Invisibly NULL.
}
\description{
A lot of historical analysis scripts rely on global objects such as variables
describing details of the sampling design or listing tree species groups or
data frames containing sampling data.
We replaced the global variables by getter functions (cf
\code{\link{get_design}}, \code{\link{get_bwi_species_groups}} and the
data frames by \pkg{bwibw} and \code{\link{get_data}}.
To stay compatible with the historical analysis scripts, this is is a wrapper
to provide those global variables and data frames to the global environment.
}
