% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_objects.R
\name{get_global_variables}
\alias{get_global_variables}
\title{Define Global Variables Needed by Historical Analysis Scripts}
\usage{
get_global_variables()
}
\value{
See \code{link{assign}}.
}
\description{
See \code{\link{get_global_objects}}.
}
\keyword{internal}
