% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_label_for_abbreviation}
\alias{get_label_for_abbreviation}
\title{Get the label for a abbreviation.}
\usage{
get_label_for_abbreviation(abbreviation, label_type = "text_label")
}
\arguments{
\item{abbreviation}{The abbreviation which shall be replaced.}

\item{label_type}{Type of the label.}
}
\value{
An expression holding a label and an approriate unit or a label as
  charater depending on the \code{label_type}. If \code{label_type} is not
  known, 'unknown name' and the unknown label type are returned as character
  vector.
}
\description{
Replace an abbreviation with the appropriate label.
}
\details{
Depending on the label type, the function calls
\code{\link{get_axis_label_for_abbreviation}} or
\code{\link{get_text_label_for_abbreviation}}. Possible label types are
'axis_label' and 'text_label' which is the default.
}
\examples{
abbreviation <- "V_DhmR"
label <- get_label_for_abbreviation(abbreviation)
print(get_label_for_abbreviation(abbreviation, label_type = 'text_label'))
label <- get_label_for_abbreviation(abbreviation, label_type = 'axis_label')
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
