% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_options}
\alias{get_options}
\title{get options from a list}
\usage{
get_options(..., name = ifelse(is.null(getOption("package_name")), "my_list",
  getOption("package_name")), remove_names = FALSE, flatten_list = TRUE)
}
\arguments{
\item{...}{see \code{\link{getOption}}/}

\item{name}{[character(1)]\cr Name of the list.}

\item{remove_names}{[boolean(1)]\cr Remove the names?}

\item{flatten_list}{[boolean(1)]\cr Return a vetcor?}
}
\value{
a (possibly named) list or a vector.
}
\description{
a convenience function for \code{\link{getOption}}.
}
\examples{
message('See help("set_options")')
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
