% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_objects.R
\name{get_package_data}
\alias{get_package_data}
\title{Load a Package's Data Into the Global Environment}
\usage{
get_package_data(package)
}
\arguments{
\item{package}{Name of the package from which to load the data.}
}
\value{
Invisibly a character vector of names of the objects loaded.
}
\description{
Load a Package's Data Into the Global Environment
}
