% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits_and_pieces.R
\name{get_period}
\alias{get_period}
\title{get period}
\usage{
get_period(label)
}
\arguments{
\item{label}{Integer which labels the period (1 or 2).}
}
\value{
time period (String).
}
\description{
gets the time period for a label.
}
\examples{
get_period(2)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
