% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_getters.R
\name{get_species_groups}
\alias{get_species_groups}
\title{get the species group list}
\usage{
get_species_groups()
}
\value{
a list with
\itemize{
    \item[\var{bagr.lab}] a vector of tree species group labels.
    \item[\var{ba.grupp}] a corresponding list of vectors giving the species
    codes for each tree species group.
    \item[\var{colors}] a corresponding vector of color definitions.
}
}
\description{
the list of species groups is used by many functions.
}
\section{warning}{
 there is a second function
\code{\link{get_bwi_species_groups}}
which does nearly the same but the return value is slightly different and on
which this functions relies. What a mess.
}

\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
