% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_text_label_for_abbreviation}
\alias{get_text_label_for_abbreviation}
\title{Get the text label for a abbreviation.}
\usage{
get_text_label_for_abbreviation(abbreviation)
}
\arguments{
\item{abbreviation}{The abbreviation which shall be replaced.}
}
\value{
The label as character. If \code{label} is not known, the
  abbreviation will be returned.
}
\description{
Replace an abbreviation with the appropriate text label.
}
\examples{
abbreviation <- "V_DhmR"
print(get_text_label_for_abbreviation(abbreviation))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
