% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{golden_ratio}
\alias{golden_ratio}
\title{calculate the golden ratio}
\usage{
golden_ratio(x)
}
\arguments{
\item{x}{the sum of the two quantities to be in the golden ratio}
}
\value{
a numeric vector of lenght 2, containing the two quantites, a being
the larger.
}
\description{
calculate the golden ratio
}
\examples{
golden_ratio(10)
}
