% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{group_district_species}
\alias{group_district_species}
\title{FIXME}
\usage{
group_district_species(groupings = list(c("Fichte", "Weißtanne", "Douglasie",
  "Kiefer", "Lärchen/sNB"), c("Buche", "Eichen"), c("Esche", "Bergahorn",
  "HBu", "sBlb", "Aln")), shares = NULL)
}
\arguments{
\item{groupings}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
