% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{grouping_species}
\alias{grouping_species}
\title{FIXME}
\usage{
grouping_species(species = c("Fichte", "Weißtanne"), minimal_share = 10,
  shares = species_shares,
  species_list = get_bwi_species_groups("regional"))
}
\arguments{
\item{species}{FIXME}

\item{minimal_share}{FIXME}

\item{shares}{FIXME}

\item{species_list}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
