% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadwood.R
\name{harmonize_deadwood}
\alias{harmonize_deadwood}
\title{Harmonizes deadwood table}
\usage{
harmonize_deadwood(deadwood, to_bwi_2 = TRUE)
}
\arguments{
\item{deadwood}{Dataframe- table with information about deadwood.}

\item{to_bwi_2}{Logical Argument, by default set to TRUE.}
}
\value{
Dataframe- table with harmonized deadwood information.
}
\description{
Function harmonizes deadwood information from BWI3 to categories of BWI2.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
