% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\docType{data}
\name{Header}
\alias{Header}
\title{batch.R}
\description{
check for batch runs, get code file names.
}
\section{Version}{
 $Id: 68c8cd12e7164e2ec25d6f4473d74fd62fbba706 $
}

\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
