% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{iVB.bilanz.bagr.akl.dkl.fun.2g}
\alias{iVB.bilanz.bagr.akl.dkl.fun.2g}
\title{Aggregiert Zuwachsbilanzkomponenten fuer eine Zuwachsperiode}
\usage{
iVB.bilanz.bagr.akl.dkl.fun.2g(baeume.vor, baeume.folg, ecken)
}
\arguments{
\item{baeume.vor}{Messdaten der Vorinventur.}

\item{baeume.folg}{Messdaten der Folgeinventur.}

\item{ecken}{Eckeninformationen.}
}
\value{
Dataframe-Tabelle mit Informationen ueber Zuwachsbilanzen, Biomasse
 in Tonnen, Grundflaeche in m² und Baumartenflaeche.
}
\description{
Funktion aggregiert Zuwachsbilanzkomponenten furer eine Zuwachsperiode aus
Aufnahmedaten der Vor- und Folgeinventur.
}
\section{Note}{
 Erwartet werden die nach Baumartengruppe-Gruppe(BaGr),
 Altersklasse (akl) und Durchmesser-Klasse (dkl) jweils zur Periodenmitte
 klassifizierten Probebaum-Datensaetze der Vor- und Folgeinventur.
}

\section{Aktualisierungen}{

 11.04.2014 (1) Einfuegen Derbholzvolumen im Hauptbestand, (2) Ausgeschiedener
   Vorrat \cr
 15.01.2015 Grundflaechenzuwachs bzw. -abgang \cr
 09.02.2015 (1) oberird. Biomasse: Beschraenkung auf Derbholzkollektiv (vom
   26.01.2015) wird beibehalten, zusaetzlich wird beim echten Einwuchs die
   Differenz  oiB2 - oiB1 gebildet! \cr
   (2) Beim Grundflaechenzuwachs wird nur das Derbholz betrachtet, daher wird
   beim echten Einwuchs die Differenz zur Derbholzschwelle gebildet:
   (BHD2^2 - 7^2)/.... Generell wird beim Derbholzkollektiv alles als Zuwachs
   betrachtet, welches innerhalb der Periode die Kriterien erf?llt:
   (a) Derbholz-Kollektiv: d.h. BHD2  > 7 (bei E- und S-Baeumen), bei den
   A-Baeumen ebenfalls (hier ist BHD1 immer >= 7) Dies gilt auch fuer die
   Biomasse, da die Veraenderung zur Vorinventur nicht erfasst wird, der
   Biomassevorrat unterhalb der Derbholzschwelle muesste abgezogen werden!
}

\section{Hinweis}{

 09.02.2015 Bei der Klassifikation nach Durchmessern zur Periodenmitte tritt
 das Phaenomen auf, dass der Teil des Kollektivs herausfaellt, dessen BHD zur
 Periodenmitte noch nicht 7 cm erreicht hat! bzw. dieser Zuwachs wird der
 Klasse BHD < 7 cm zugewiesen.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
