% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{iVB.bilanz.bagr.akl.dkl.fun}
\alias{iVB.bilanz.bagr.akl.dkl.fun}
\title{Aggregiert Zuwachsbilanzkomponenten fuer eine Zuwachsperiode}
\usage{
iVB.bilanz.bagr.akl.dkl.fun(baeume.vor, baeume.folg, ecken)
}
\arguments{
\item{baeume.vor}{Messdaten der Vorinventur.}

\item{baeume.folg}{Messdaten der Folgeinventur.}

\item{ecken}{Eckeninformationen.}
}
\value{
Dataframe-Tabelle mit Informationen ueber Zuwachsbilanzen, Biomasse
 in Tonnen, Grundflaeche in m² und Baumartenflaeche.
}
\description{
Funktion aggregiert Zuwachsbilanzkomponenten furer eine Zuwachsperiode aus
Aufnahmedaten der Vor- und Folgeinventur.
}
\section{Note}{
 Erwartet werden die nach Baumartengruppe-Gruppe(BaGr),
 Altersklasse (akl) und Durchmesser-Klasse (dkl) jweils zur Periodenmitte
 klassifizierten Probebaum-Datensaetze der Vor- und Folgeinventur.
}

\section{Aktualisierungen}{

 11.04.2014 (1) Einfuegen Derbholzvolumen im Hauptbestand, (2) Ausgeschiedener
   Vorrat \cr
 09.02.2015 Biomassezuwachs auf Derbholz-Kollektiv beschraenkt (echter
 Einwuchs);  Grund: Einwuchs ins Derbholz-Kollektiv wird als Zuwachs explizit
 erfasst, da bei Vorinventur nicht vorhanden. Wuerde auch die Biomasse
 unterhalb der Derbholzschwelle einbebzogen, muesste der Biomasse-Vorrat
 unterhalb der Derbholzschwelle der Vorinventur abgezogen werden! Beim echten
 Einwuchs wird die Differenz gebildet oiB2 -oiB1)...
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
