% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{iVB.ew.bagrupp.akl.dkl.stratum.fun.2g}
\alias{iVB.ew.bagrupp.akl.dkl.stratum.fun.2g}
\title{Aggregiert Daten fuer Zuwachs und ausgeschiedenen Vorrat}
\usage{
iVB.ew.bagrupp.akl.dkl.stratum.fun.2g(baeume.23, baeume.3, BA.grupp, A.klass,
  D.klass, auswahl, A, trakte.3 = get_data("trakte.3"),
  ecken.3 = get_data("ecken.3"), ecken.2 = get_data("ecken.2"))
}
\arguments{
\item{baeume.23}{Tabelle mit Baumdaten aus BWI 2.}

\item{baeume.3}{Tabbelle mit Baumdaten aus BWI 3.}

\item{BA.grupp}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199),c(200:299))).}

\item{A.klass}{Liste mit den Klassifizierungsparametern fuers Alter: z.B.
list(A.ob=160, A.b=20).}

\item{D.klass}{Liste mit den Klassifizierungsparametern fuer Durchmesser z.B.
list(D.unt=0, D.ob=70, D.b=10, Ndh=T), Ndh (Nicht-Derbholz) = T bedeutet,
dass zusaetzlich Nicht-Dh (unter 7 cm) ausgewiesen wird, sonst gilt
\code{D.unt} als unterste Schwelle.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}

\item{trakte.3}{Trakt information.}

\item{ecken.3}{Ecken information.}

\item{ecken.2}{Ecken information.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit Datum und
 genutzter Baumversion), \strong{Stratum} (\code{auswahl}), \strong{HBF}
 (Holzbodenflaeche in ha), \strong{se.HBF} (Standardfehler Holzbodenflaeche),
 \strong{Attribute} (Vektor mit berechneten Attributen), \strong{Groessen}
 (vektor mit berechneten Grroessen ("Wert", "Standardfehler")), \strong{BAGR}
 (Labels fuer Baumartengruppen aus \code{ba.grupp}), \strong{AKL} (Labels der
 Altersklassen), \strong{DKL} (Labels der Durchmesserklassen),
 \strong{iVB.bagr.akl.dkl} (Zuwachs-Tabelle), \strong{VB.A.bagr.akl.dkl}
 (Tabelle mit ausgeschiedenem Vorrat).
}
\description{
Funktion aggregiert Daten fuer Zuwachs und ausgeschiedenen Vorrat von BWI 2
zu BWI 3 nach frei definierbaren Baumarten-, Altersklassen und
Durchmesserklassengruppen.
}
\section{Aktualisierungen}{

 24.01.2015 Erweiterung um Grundflaechenzuwachs \cr
 27.01.2015 Korrektur betreffend Aggregation auf alle Baumarten durch eigene
 seperate Aggregation auf Baumartengrupe "AlleBA" \cr
 15.12.2014 Erweiterung um "AlleBA" \cr
 23.04.2014 Uebergabe von \code{baeume.23} und \code{baeume.3} \cr
 30.07.2014 kleinere Korrekturen: Version mit Berechnung des jaehrlichen
 flaechenbezogenen Zuwachses, wobei der jaehrliche Gesamtzuwachs als Y im
 Ratio-Schaetzer steht (entspricht Vorgehen des TI): \code{lz=SumjVZ/SumBAF},
 wobei \code{SumjVZ=Sum(VZ.i/PL.i)}. Alternative: erweiterter
 Ratio-Schaetzer: \code{lZ = SumVZ/SumBAF/mPL}.
}

\section{Grundsatz}{
 Der Periodenzuwachs wird im sog. gemeinsamen Netz
 (= Schnittmenge) einer Befundeinheit (Stratum, Domaene) berechnet benoetigte
 Tabellen: <baeume.3>, <baeume.23>, <ecken.3>, <ecken.2>, <trakte.3>.
}

\section{Hinweis}{
 "Endwert-Verfahren" bzgl. S- und E-Baeume, Bilanzierung
 inkomaptibel mit Vorratsdifferenz aus Zustandsgroessen! \cr
 Echter Einwuchs wird getrennt ausgewiesen, ausschliesslich nach dem Kriterium
 (modellierter) BHD1 < 7 cm (abweichend von bisherigem Kriterium Entf < 1.75
 & BHD2 < 22).
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
