% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{map_abbreviations_to_labels}
\alias{map_abbreviations_to_labels}
\title{Get labels for different abbreviations.}
\usage{
map_abbreviations_to_labels(x)
}
\arguments{
\item{x}{A factor or vector containing abbreviations.}
}
\value{
A factor or vector containing labels.
}
\description{
Replace abbreviations with the appropriate text labels.
}
\examples{
x <- c("V_DhmR", "DKL", "MPW")
labels <- map_abbreviations_to_labels(x)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
