% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{map_labels_to_abbreviations}
\alias{map_labels_to_abbreviations}
\title{Get abbreviations of different labels.}
\usage{
map_labels_to_abbreviations(x)
}
\arguments{
\item{x}{A factor or vector containing labels.}
}
\value{
A factor or vector containing abbreviations.
}
\description{
Replace labels with the appropriate abbreviations.
}
\examples{
x <- c("Derbholzvorrat", "Durchmesserklasse", "Mittlerer Privatwald")
abbreviations <- map_labels_to_abbreviations(x)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
