% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{mbaf.bagr.alt.bhd.pm.fun}
\alias{mbaf.bagr.alt.bhd.pm.fun}
\title{Klassifiziert und aggegiert die mittlere Baumartenflaeche}
\usage{
mbaf.bagr.alt.bhd.pm.fun(baeume.vor, baeume.folg, A.klass, D.klass)
}
\arguments{
\item{baeume.vor}{Tabelle fuer Vorinventur mit Baumartengruppe,
Attribute: tnr, enr, bnr, pk, pl, ba, bagr, bhd1, bhd2, alt1, alt2, stfl1,
stfl2.}

\item{baeume.folg}{Tabelle fuer Folgeinventur mit Baumartengruppe,
Attribute: tnr, enr, bnr, pk, pl, ba, bagr, bhd1, bhd2, alt1, alt2,
stfl1, stfl2.}

\item{A.klass}{Liste mit den Klassifizierungsparametern fuers Alter: z.B.
list(A.ob=160,A.b=20).}

\item{D.klass}{Liste mit den Klassifizierungsparametern fuer Durchmesser z.B.
list(D.unt=0,D.ob=70,D.b=10,Ndh=T), Ndh (Nicht-Derbholz) = T bedeutet, dass
zusaetzlich Nicht-Dh (unter 7 cm) ausgewiesen wird, sonst gilt \code{D.unt}
als unterste Schwelle.}
}
\value{
Dataframe-Tabelle mit folgenden Attributen: tnr, bagr, akl.pm,
 dkl.pm, baf1, baf2, mbaf, mbaf.hb. Wobei mbaf die mittlere Baumartenflaeche
 und mbaf.hb die mittlere Baumartenflaeche im Hauptbestand enthaelt.
}
\description{
Funktion fuehrt eine Klassifikation der mittleren Baumartenflaeche nach
Baumarten-Gruppe, BHD und Alter zur Periodenmitte durch und aggregiert nach
Trakt.
}
\section{Erstellungsdatum}{
 04.03.2014.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
