% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{melt_species_attribute_group}
\alias{melt_species_attribute_group}
\title{Predicts values for species grouped by a variable and BWI}
\usage{
melt_species_attribute_group(list, species, abbreviation, group)
}
\arguments{
\item{list}{Output (List) of BWI3_HR_Funktionen (e.g.
\code{\link{FVBN.bagrupp.akl.dkl.stratum.fun.2a}}).}

\item{species}{String which specifies the tree species for which calculation
shall be done. Possible inputs: FI TA DGL KI LAE BU EI ALH ALN.}

\item{abbreviation}{String which speciefies the value for which the
calculation shall be done. Possible input depends on the existing values in
\code{list}.}

\item{group}{String which specifies the value by which the calculation shall
be grouped.}
}
\value{
Data frame with prediction for the given \code{abbreviation} and
 \code{species}, grouped by \code{group} and BWI.
}
\description{
Function predicts a value which is speciefied by the user for a variable
species. The prediction is done grouped by the different BWI's and a grouping
variable, given by the user.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
