% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_species_shares.R
\name{melt_tree_species_percent}
\alias{melt_tree_species_percent}
\title{Calculates relative tree species area}
\usage{
melt_tree_species_percent(statistics_list, bagr = get_bwi_species_groups())
}
\arguments{
\item{statistics_list}{String which defines the list holding the statistical
data.}

\item{bagr}{A list setting up tree species groups. Stick with the default.}
}
\value{
Data Frame with information about relative tree species area for each
 species. Information for all three BWI's.
}
\description{
Function calculates the relative tree species area for each tree species
group for all three BWI's.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
