% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{ntns.stratum.fun}
\alias{ntns.stratum.fun}
\title{Berechnet Flaechen und Flaechenanteile der Naturnaehestufen}
\usage{
ntns.stratum.fun(ntns.te, ecken, trakte, A, auswahl, bwi, natwg)
}
\arguments{
\item{ntns.te}{alle Traktecken die Naturnaehestufen in allen Varianten
(Schichten, Gesamtwert fuer Hauptbestockung) fuer BWI 2 und 3 sowie fuerr BWI
2 in den 2 Vaianten natWG BWI 2 und natWG BWI 3.}

\item{ecken}{Stichprobenmerkmale der jeweiligen BWI (2 oder 3).}

\item{trakte}{Traktmermale (m, m_HB, m_bHB, m_Wa).}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{bwi}{Angabe welche BWI (2 oder 3).}

\item{natwg}{Differenzierung bei Auswertung fuer BWI 2: 2: nat. WG der 3
damaligen BWI 2 oder 3: bei BWI 3 aktualisierte nat. WG, die rueckwirkend
auf BWI 2 angewandt wird.}
}
\value{
Liste mit folgenden Komponenten: \strong{Datum}, \strong{HBF}
 (Holzbodenflaeche in ha), \strong{se.HBF} (Standardfehler der
 Holzbodenflaeche), \strong{NTNS} (Vektor mit den verschiedenen ntns),
 \strong{NTNS.Flaeche.Anteil} (absoluter und relativer Flaechenanteil vom
 Inventurgebiet (?) und jeweiliger Standardfehler).
}
\description{
Funktion berechnet fuer das in \code{auswahl} definierte Stratum die Flaechen
und Flaechenanteile (\%) der 5 Naturnaehestufen (ntns) fuer BWI 2 oder BWI 3
}
\section{Erstellungsdatum}{
 19.08.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
