% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{plot_by_group}
\alias{plot_by_group}
\title{Plots data by group}
\usage{
plot_by_group(data, subs, x = "Baumartengruppe")
}
\arguments{
\item{data}{Output table of \code{\link{all_data_by_group}}}

\item{subs}{String which defines subset parameters, e.g. "Eigentumsart ==
'Privatwald'".}

\item{x}{String which defines the column in data which shall be used for the
grouping on the x-axis. by default "Baumartengruppe"}
}
\value{
Bar plot
}
\description{
Function creates a bar diagramm plot of the output of
\code{\link{all_data_by_group}}. The predicted value for the abbreviation is
used to define the height of the bars. The bars are grouped by
"Baumartengruppe" (by default) or another parameter.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
