% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_growth_loss}
\alias{plot_growth_loss}
\title{FIXME}
\usage{
plot_growth_loss(g_l_12, g_l_23, g_l_12_all, do_errors_relative = FALSE,
  graphic_width = get_options("graphics_width"),
  graphic_height = graphics_height, graphic_directory = graphics_directory,
  file_name_district = regional_file_name,
  title_district = krs.grupp$string[i],
  species_groups_labels = district_groups$ba.text)
}
\arguments{
\item{g_l_12}{FIXME}

\item{g_l_23}{FIXME}

\item{g_l_12_all}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{title_district}{FIXME}

\item{species_groups_labels}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
