% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_ownership}
\alias{plot_ownership}
\title{FIXME}
\usage{
plot_ownership(ownerships, graphic_width = get_options("graphics_width"),
  graphic_height = graphics_height, graphic_directory = graphics_directory,
  file_name_district = regional_file_name,
  title_district = krs.grupp$string[i])
}
\arguments{
\item{ownerships}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{title_district}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
