% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_silviculturally_relevant}
\alias{plot_silviculturally_relevant}
\title{FIXME}
\usage{
plot_silviculturally_relevant(relevant_species_2, relevant_species_3,
  do_errors_relative = FALSE, graphic_width = get_options("graphics_width"),
  graphic_height = graphics_height, graphic_directory = graphics_directory,
  file_name_district = regional_file_name,
  title_district = krs.grupp$string[i],
  species_groups_labels = c(district_groups$ba.text, "Alle BA"),
  have_title = FALSE)
}
\arguments{
\item{relevant_species_2}{FIXME}

\item{relevant_species_3}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{title_district}{FIXME}

\item{species_groups_labels}{FIXME}

\item{have_title}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
