% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{plot_species_attribute_group}
\alias{plot_species_attribute_group}
\title{Plots the output of \code{\link{melt_species_attribute_group}}}
\usage{
plot_species_attribute_group(data)
}
\arguments{
\item{data}{Output Dataframe of \code{\link{melt_species_attribute_group}}.}
}
\value{
Plot which shows calculation of
 \code{\link{melt_species_attribute_group}}.
}
\description{
Function plots the output of \code{\link{melt_species_attribute_group}}.
Different BWI's are shown in three different green shades. The grouping value
is shown on the x axis and the prediction is the y value.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
