% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{prettify_data_frame}
\alias{prettify_data_frame}
\title{apply prettyNum to a data.frame, preserve its row.names}
\usage{
prettify_data_frame(data_frame, digits = 0)
}
\arguments{
\item{data_frame}{The data which to prettyNum().}

\item{digits}{The number of digits to be rounded to.}
}
\value{
the modified data.frame.
}
\description{
apply prettyNum to a data.frame, preserve its row.names
}
\examples{
prettify_data_frame(utils::head(airquality * 10^4))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
