% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadwood.R
\name{purge_deadwood_to_2}
\alias{purge_deadwood_to_2}
\title{Selects a subset from deadwood table}
\usage{
purge_deadwood_to_2(totholz)
}
\arguments{
\item{totholz}{Dataframe- table with information about deadwood.}
}
\value{
Subset of input table.
}
\description{
Function selects a subset from the deadwood table by following condition:
\code{tart != 4 & tbd >= 20) | (tart == 4 & (tbd >= 60 | lge >= 0.5)}.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
