% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{r.variance.fun}
\alias{r.variance.fun}
\title{Berechnet Ratio-Schaetzer und seine Varianz}
\usage{
r.variance.fun(xy, n)
}
\arguments{
\item{xy}{Matrix mit  x und y als Spalten.}

\item{n}{Anzahl der Spalten der Matrix, Anzahl der Trakte im Inventurgebiet.}
}
\value{
Liste mit Ratio-Schaetzer (\code{R.xy}) und Varianz des Ratio-
 Schaetzers (\code{V.R.xy}).
}
\description{
Funktion berechnet Ratio-Schaetzer und seine Varianz ueber Fehlerfortpflanzung.
}
\section{Erstellungsdatum}{
 13.03.2014
}

\section{Note}{

 Version mit \code{xy}, das grundsaetzlich alle Trakte des
 Inventurgebiets umfasst. \cr
 Der zugehoerige Ratio-Schaetzer ist R = y/x = S.y/n/S.x/n \cr
 <trakte> ist der Datensatz der Trakte im gesamten Inventurgebiet, die Trakte
 sind die primaeren Stichprobeneinheiten, <trakte> enthaelt (u.a.) die Anzahl
 <m> der Traktecken (Subplots) je Trakt --> nicht im Code enthalten (?) \cr
 \code{y} sind die Beobachtungen im Stratum (der Untermenge der Trakte), \cr
 \code{x}: sind (1) bei Totalen die  Traktecken (Wald/Nichtwald), (2) bei
 Ratio zwischen beob. Groessen die jeweilige Groesse im Stratum (als Untermnege
 der Trakte im Inventurgebiet).
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
