% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{r.xxy.variance.fun}
\alias{r.xxy.variance.fun}
\title{Berechnet einen "kombinierten" Ratio-Schaetzer und seine Varianz}
\usage{
r.xxy.variance.fun(xy, n)
}
\arguments{
\item{xy}{Matrix mit  x1, x2 und y als Spalten.}

\item{n}{Anzahl der Spalten der Matrix, Anzahl der Trakte im Inventurgebiet.}
}
\value{
Liste mit Ratio-Schaetzer(\code{R.xxy}) und Varianz des Ratio-
 Schaetzers (\code{V.R.xxy}).
}
\description{
Funktion berechnet einen "kombinierten" Ratio-Schaetzer mit 2 x-Groessen im
Nenner und seine Varianz ueber Fehlerfortpflanzung.
}
\section{Erstellungsdatum}{
 10.03.2014
}

\section{Note}{
 version mit \code{xy}, das grundsaetzlich alle Trakte des
 Inventurgebiets umfasst. \cr
 Der zugehoerige Ratio-Schaetzer ist R = y/x = S.y/n/S.x1/n * S.t/S.x2. \cr
 \code{t} ist eine Indikatorvariable, welche das Auftreten einer Beobachtung
 von x2 (x2 > 0) erfasst (t = 1, wenn x2 > 0; 0, sonst) \cr
 \code{y} sind die Beobachtungen im Stratum (der Untermenge der Trakte),
 x1: sind (1) bei Totalen die  Traktecken (Wald/Nichtwald), (2) bei Ratio
 zwischen beob. Groessen die jeweilige Groesse im Stratum (als Untermnege der
 Trakte im Inventurgebiet).
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
