% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{reassemble_by_group}
\alias{reassemble_by_group}
\title{Reassembles by ownership and a variable abbreviation}
\usage{
reassemble_by_group(owner, abbreviation)
}
\arguments{
\item{owner}{String which defines the owner.}

\item{abbreviation}{String which defines the variable which shall be
predicted.}
}
\value{
Dataframe.
}
\description{
Fix me! What do I do?
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
