% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{regeneration}
\alias{regeneration}
\title{FIXME}
\usage{
regeneration(r2_a, r2_bs1, r2_bs2, r3_a, r3_bs1, r3_bs2,
  graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii)
}
\arguments{
\item{r2_a}{FIXME}

\item{r2_bs1}{FIXME}

\item{r2_bs2}{FIXME}

\item{r3_a}{FIXME}

\item{r3_bs1}{FIXME}

\item{r3_bs2}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
