% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{revalue_data}
\alias{revalue_data}
\title{Revalue abbreviations to labels.}
\usage{
revalue_data(datacolumn)
}
\arguments{
\item{datacolumn}{Datacolumn of a table containing abbreviations.}
}
\value{
invisible(status) Which becomes TRUE when the abbrevations were
  successfully replaced.
}
\description{
Replace abbreviations with the appropriate text labels.
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
