% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{round_and_prettify}
\alias{round_and_prettify}
\title{round() and prettyNum() numbers}
\usage{
round_and_prettify(x, digits = 0, big_mark = ",", decimal_mark = ".")
}
\arguments{
\item{x}{the numbers to be modified.}

\item{digits}{The digits passed to round().}

\item{big_mark}{the big.mark passed to \code{prettyNum}.}

\item{decimal_mark}{the decimal.mark passed to \code{prettyNum}.}
}
\value{
the modified numbers.
}
\description{
The majority of readers can't cope with numbers like 12000.01. So we
change it to 12,000.
}
\examples{
round_and_prettify(airquality[1:10, "Ozone"]*10^4+0.1)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
