% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{round_and_prettify_german}
\alias{round_and_prettify_german}
\title{round() and prettyNum() numbers}
\usage{
round_and_prettify_german(x, digits = 0)
}
\arguments{
\item{x}{the numbers to be modified.}

\item{digits}{The digits passed to round().}
}
\value{
the modified numbers.
}
\description{
The majority of \emph{german} readers can't cope with numbers like 12000.01.
So we change it to 12.000.
}
\note{
This is just an alias for
\code{\link{round_and_prettify}}(\ldots, big_mark = '.')
}
\examples{
round_and_prettify_german(airquality[1:10, "Ozone"]*10^4+0.1)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
