% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits_and_pieces.R
\name{set_to_limit}
\alias{set_to_limit}
\title{limit the values in a data.frame's column to a limit}
\usage{
set_to_limit(data_frame, condition)
}
\arguments{
\item{data_frame}{Give the data.frame.}

\item{condition}{See Details}
}
\value{
the modified data.frame
}
\description{
without changing the original data.frame
}
\section{Details}{

\code{condition} of form \emph{variable operator limit} and will set all
values of \emph{variable} to \emph{limit} where the condition is TRUE.
lower than Zero to Zero.
}

\examples{
set_to_limit(mtcars, mpg < 20)
set_to_limit(set_to_limit(mtcars, mpg < 20), mpg > 23)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
