% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{stamm.merkmale.bagr.akl.fun}
\alias{stamm.merkmale.bagr.akl.fun}
\title{Funktion erzeugt aus Ergebnistabelle der Funktion \code{\link{verjg.bagr.fun}}
bzw. \code{\link{verjg.bagrupp.fun}} eine Dataframe-Tabelle.}
\usage{
stamm.merkmale.bagr.akl.fun(wzp4.merkmale, baeume, ecken, trakte, A, auswahl,
  merkmale, ba.grupp, A.klass, trakte.3 = get_data("trakte.3"))
}
\arguments{
\item{wzp4.merkmale}{Eine Tabelle der besonderen Stammmerkmale,
BWI 3
merkmal.list.3 <- c("tot","jschael","aeschael","ruecke","pilz","harz","kaefer",
               "sstamm","faulkon","hoehle","bizarr","uralt","horst","mbiotop")
BWI 2
merkmal.list.2 <- c("tot","jschael","aeschael","ruecke","pilz","harz","kaefer",
              "sstamm","hoehle")}

\item{baeume}{Datentabelle mit Bauminformationen. Die Tabelle muss mindestens
die Attribute TNr, ENr, BA, Alt., BHD., VolV., oiB., NHa., StFl. enthalten
(. steht fuer 1 oder 2, je nachdem welcher Zustand ausgewertet werden soll),
man kann auch die selektierten Attribute mit Namen ohne Kennziffer
uebergeben, wenn bereits eine eindeutige Auswahl der Attribute in
\code{baeume} uebergeben wird.}

\item{ecken}{Eckenmerkmale.}

\item{trakte}{Traktmerkmale.}

\item{A}{Gesamtflaeche in ha des Inventurgebiets zum jeweiligen
Inventurzeitpunkt (sollte eigentlich konstant sein).}

\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.:
list(Wa=c(3,5), Begehbar=1).}

\item{merkmale}{legt das/die Merkmal/e fest (es koennen mehrere angegeben
werden, diese werden als Vereinigungsmenge je Baum interpretiert!)}

\item{ba.grupp}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199), c(200:299))).}

\item{A.klass}{Liste mit den Klassifizierungsparametern fuers Alter: z.B.
list(A.ob=160, A.b=20).}

\item{trakte.3}{Trakt information.}
}
\value{
Eine Liste mit Zeug
}
\description{
Funktion erzeugt aus Ergebnistabelle der Funktion \code{\link{verjg.bagr.fun}}
bzw. \code{\link{verjg.bagrupp.fun}} eine Dataframe-Tabelle.
}
\note{
da nur das WZP-4-Kollektiv beruecksichtigt wird, koennen keine
Baumartenflaechen berechnet werden!
}
\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
