% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{stamm.merkmale.bagr.fun}
\alias{stamm.merkmale.bagr.fun}
\title{Aggregiert Flaechenwerte und Standardfehler fuer gegebene Stammmerkmale}
\usage{
stamm.merkmale.bagr.fun(wzp4.merkmale, baeume, ecken, trakte, A, auswahl,
  merkmale, ba.grupp, trakte.3 = get_data("trakte.3"))
}
\arguments{
\item{wzp4.merkmale}{Tabelle, welche die besonderen Stammmerkmale enthaelt;
moegliche Merkmale sind: \cr
BWI 3: merkmal.list.3 <- c("tot", "jschael", "aeschael", "ruecke", "pilz",
"harz", "kaefer", "sstamm", "faulkon", "hoehle", "bizarr", "uralt", "horst",
"mbiotop"). \cr
BWI 2: merkmal.list.2 <- c("tot", "jschael", "aeschael", "ruecke", "pilz",
"harz", "kaefer", "sstamm", "hoehle").}

\item{baeume}{Probebaum-Attribute.}

\item{ecken}{Trakteckenmerkmale.}

\item{trakte}{traktbezogene Informationen (u.a. m, m_HB, ...).}

\item{A}{Flaeche des Inventurgebietes in ha.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{merkmale}{legt das/die Merkmal/e fest (es koennen mehrere angegeben
werden, diese werden als Vereinigungsmenge je Baum interpretiert!)}

\item{ba.grupp}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab"). Z.B. list(bagr.lab = c("FiTa",
"DglKiLae","Bu","Ei","BLb","WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199),c(200:299))).}

\item{trakte.3}{Trakt information.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und genutzten Baeumen(?)), \strong{Stratum}
 (\code{auswahl}), \strong{nTE} (Anzahl Ecken im Stratum), \strong{HBF}
 (Holzbodenflaeche in ha), \strong{se.HBF} (Standardfehler HBF), \strong{BL}
 (Flaeche der Bloessen in ha), \strong{se.BL} (Standardfehler BL),
 \strong{iBL} (Flaeche der idellen Bloessen ("Luecken") in ha),
 \strong{se.iBL} (Standardfehler iBL), \strong{LK} (relative
 Lueckenkorrektur), \strong{se.LK}  (Standardfehler LK), \strong{Merkmal} (?),
 \strong{BAGR} (Baumgruppenlabel der eingegebenen Liste \code{ba.grupp}),
 \strong{Kennwerte} (Vektor mit Kennwerten), \strong{Kennwert_Tabelle_BAGR}
 (2-D array mit Werten und Standardfehler fuer gesamte Stueckzahl, Stueckzahl je ha, gesamtes Volumen,
 Volumen je ha, Stueckvolumen, gesamte oberirdische Biomasse, oberirdische
 Biomasse je ha, Anteil Merkmal an Gesamtflaeche, Anteil des Merkmal-Vorrats
 am Gesamtvorrat).
}
\description{
Funktion agregiert Flaechenwerte und Standardfehler fuer Gesamtwerte und
Werte je ha fuer ein definiertes Stratum. Dabei werden Flaechen von Waldecken
berechnet und Stammwerte zu Stueckzahl, Volumen und oberirdische Biomasse.
}
\section{Version}{
 1.0 vom 04.11.2014
}

\section{Hinweis}{

 nur fuer BWI 2 und 3! \cr
 Da nur das WZP-4-Kollektiv beruecksichtigt wird, koennen keine
 Baumartenflaechen berechnet werden!
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
