% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{stratum.fun}
\alias{stratum.fun}
\title{Selektiert die zu einem Stratum gehoerenden Stichproben}
\usage{
stratum.fun(auswahl, ecken)
}
\arguments{
\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl erfolgt.}

\item{ecken}{Tabelle mit allen zur Selektion dienenden Eckenmerkmalen.}
}
\value{
Untermenge der Ecken, die der Auswahl entsprechen.
}
\description{
"Stratifikator"-Funktion: selektiert die zu einem Stratum (Befundeinheit,
Subdomaene, Subpopulation) gehoerenden Stichproben (Traktecken) anhand
verschiedener Merkmale, die in der Tabelle \code{ecken} zusammengefasst sind
die Tabelle \code{ecken} enthaelt nur die Stichproben im Wald (Holzboden,
Nichtholzboden, begehbar, nicht begehbar). Aufnahmen liegen natuerlich nur
fuer die Traktecken auf begehbarem Holzboden vor.
}
\section{Erstellungsdatum}{
 20.09.2013
}

\section{Aktualisierungen}{
 01.12.2014 leere Menge korrigiert
}

\section{Note}{
 in \code{stratum} wird zunaechst die Gesamt-Menge, also die
 vollstaendige Tabelle \code{ecken} uebergeben, die dann entsprechend der in
 der Liste \code{auswahl} uebergebenen Attribute und den zulaessigen
 Werte-Mengen als Untermenge heraus gefiltert wird.
}

\section{Refactored the whole thing Fri Aug 12 15}{
33:04 CEST 2016 by
\email{dominik.cullmann@forst.bwl.de}.
 See commit 57504d4e9570bb560c38195d3f308617c72c1d06.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
