% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_things.R
\name{tex_umlauts}
\alias{tex_umlauts}
\title{Tex code for german umlauts}
\usage{
tex_umlauts(string)
}
\arguments{
\item{string}{Non ASCII code for german umlaut.}
}
\value{
Tex code for german umlaut.
}
\description{
Gets tex code for german umlauts.
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
