% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R, R/regional.R
\name{theme_bold_legend}
\alias{theme_bold_legend}
\alias{theme_bold_legend}
\title{Sets plot legend bold}
\usage{
theme_bold_legend()

theme_bold_legend()
}
\value{
theme which sets legend of a plot bold.
}
\description{
Function produces a theme which sets the legend of a plot bold (black
colour and letter size 16).

FIXME
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
\keyword{internal}
