% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R, R/regional.R
\name{theme_bold_title}
\alias{theme_bold_title}
\alias{theme_bold_title}
\title{Sets plot title bold}
\usage{
theme_bold_title()

theme_bold_title()
}
\value{
theme which sets axes of a plot bold.
}
\description{
Function produces a theme which sets the title of a plot bold (black colour
and letter size 20).

FIXME
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
\keyword{internal}
