% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R, R/throw.R
\name{throw}
\alias{throw}
\alias{throw}
\title{throw a condition}
\usage{
throw(message_string, system_call = sys.call(-1), ...)

throw(message_string, system_call = sys.call(-1), ...)
}
\arguments{
\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}

\item{...}{Arguments to be passed to \code{\link{structure}}.}

\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}

\item{...}{Arguments to be passed to
\code{\link[base:structure]{base::structure}}.}
}
\value{
FALSE. But it doesn't return anything, it stops with a
condition of class c("bundeswaldinventur", "error", "condition").

The function does never return anything, it stops with a
condition of class c("error", "bundeswaldinventur", "condition").
}
\description{
throws a condition of class c("bundeswaldinventur", "error", "condition").

Throws a condition of class c("error", "bundeswaldinventur", "condition").
}
\details{
We use this condition as an error dedictated to \pkg{bundeswaldinventur}.

We use this condition as an error dedicated to \pkg{ bundeswaldinventur.}
}
\examples{
tryCatch(bundeswaldinventur:::throw("Hello error!"),
         bundeswaldinventur = function(e) return(e))
tryCatch(bundeswaldinventur:::throw("Hello error!"),
         bundeswaldinventur = function(e) return(e))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
\keyword{internal}
