% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{to_tex}
\alias{to_tex}
\title{FIXME}
\usage{
to_tex(..., file_name = get("dot_district_tex"), append = TRUE)
}
\arguments{
\item{...}{FIXME}

\item{file_name}{FIXME}

\item{append}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
