% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR_Totholz.R
\name{Totholz.bagr.art.zg.stratum.fun}
\alias{Totholz.bagr.art.zg.stratum.fun}
\title{Funktion aggregiert Totholz-Volumen (Maximalvariante)}
\usage{
Totholz.bagr.art.zg.stratum.fun(totholz, ecken, trakte, bwi, A, D.klass,
  auswahl)
}
\arguments{
\item{totholz}{Dataframe-Tabelle mit Totholz-Informationen. Die Tabele muss
mindestens folgende Attribute enthalten: Tnr, Enr, Nr, Tbagr, Tart, Tzg,
Tbd, Tsd,  Tl, Tvol, Anz, Thf. Hinweis: BWI3 -> Durchmesser: Tbd, Tsd;
Laenge: Tl. BWI2 -> Durchmesser: Dm; Laenge: Lge.}

\item{ecken}{Dataframe-Tabelle mit Eckenmerkmalen.}

\item{trakte}{Dataframe-tabelle mit Traktmerkmalen.}

\item{bwi}{Angabe zu BWI- Nummer (2 oder 3), um mit denselben Algorithmen
fuer beide Zustaende arbeiten zu koennen.}

\item{A}{Flaeche in ha des Inventurgebietes zum jeweiligen Inventurzeitpunkt
(sollte konstant sein).}

\item{D.klass}{Liste mit den Klassifizierungsparametern fuer Durchmesser z.B.
list(D.unt=0, D.ob=70, D.b=10, Ndh=T), Ndh (Nicht-Derbholz) = T bedeutet,
dass zusaetzlich Nicht-Dh (unter 7 cm) ausgewiesen wird, sonst gilt
\code{D.unt} als unterste Schwelle.}

\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum ueber die Funktion
\code{\link{stratum.fun}} erfolgt. Die Funktion benoetigt des weiteren die
Tabelle <bacode>. Bsp.: list(Wa=c(3,5), Begehbar=1).}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und version.totholz.b), \strong{Stratum} (\code{auswahl}),
 \strong{nTE} (Anzahl der Ecken im Stratum), \strong{HBF} (Holzbodenflaeche),
 \strong{se.HBF} (Standardfehler der Holzbodenflaeche), \strong{BAGR.Liste}
 (Vektor mit Baumartengruppen), \strong{DKL} (Labels der Durchmesserklassen),
 \strong{T.ThVN.Bagr.Tart.Tzg.Dkl} (Array mit Volumenwerten, Gesamtanzahl und
 Standardfehler nach Baumartengruppen, Totholzart, Zersetzungsgrad und
 Durchmesserklassen), \strong{ThVN.ha.Bagr.Tart.Tzg.Dkl} (Array mit
 Volumenwerten, Gesamtanzahl je ha und Standardfehler nach Baumartengruppen,
 Totholzart, Zersetzungsgrad und Durchmesserklassen),
 \strong{nT.Bagr.Tart.Tzg.Dkl} (Array mit Trakteckenanzahl nach
 Baumartengruppen, Totholzart, Zersetzungsgrad und Durchmesserklassen)
}
\description{
Funktion wertet nach Totholz-BA-Gruppen \code{Tbagr}, Totholzart \code{Tart}
und Totholzzersetzungsgrad \code{Tzg} im Stratum aus. Verallgemeinerte
Version fuer die Auswertung unterschiedlicher Inventurzeitpunkte, also BWI 2
und BWI 3.
}
\section{version}{
 1.0 (kae/17.02.2014)
}

\section{Hinweis}{
 Auf Grund verschiedener Inventurzeitpunkte werden die
 Datentabellen \code{totholz}, die Eckenmerkmale \code{ecken} und die
 \code{trakte} als Argumente uebergeben. \cr
 "Harmonisieren" der Totholzaufnahme BWI 2 und 3 -> Basis-Attribute (BWI 2):
 <TBAGr>, <TArt>, <TZg>, <TVol>, <Dm>, <Lge>, <THf>. Neu bei der BWI 3 <Tbd>:
 Durchmesser am staerkeren Ende bzw. BHD und <Tsd>: Durchmesser am duennen
 Ende (nur bei Tart 1 und 13). Fuer die Klassifikation relevant ist bei
 Bruchstuecken der Mitten-Durchmesser, bei Stuecken mit Wurzelanlauf (stehend
 oder liegend) ist es der BHD. Mittendurchmesser wird berechnet, wenn
 Tsd > 0, aus Volumen und Laenge: <Dm> = sqrt(tvol/tl/pi)*200; die
 Bezeichnung <Dm> entspricht auch derjenigen der BWI 2.Wegen der
 Einheitlichkeit mit BWI 2 wird der Bezeichner <tl> in <lge> umgewandelt. \cr
 Vergleichbarkeit mit BWI 2 -> Wenn(([Tart]=4 Und ([Tbd]>=60 Oder [Tl]>=0,5))
 Oder ([Tart]<>4 Und [Tbd]>=20). \cr
 Um Konflikte mit unterschiedlicher Gross-/Kleinschreibung bei den
 Attributnamen zu vermeiden, werden innerhalb dieser Funktion alle
 Attributnamen auf Kleinschreibung umgestellt.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
