% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR_Totholz.R
\name{Totholz.klass.stratum.fun}
\alias{Totholz.klass.stratum.fun}
\title{Funktion aggregiert Totholz-Volumen nach Klassifikation}
\usage{
Totholz.klass.stratum.fun(totholz, ecken, trakte, bwi, krit, klass, A, auswahl,
  strict = FALSE)
}
\arguments{
\item{totholz}{Dataframe-Tabelle mit Totholz-Informationen. Die Tabele muss
mindestens folgende Attribute enthalten: Tnr, Enr, Nr, Tbagr, Tart, Tzg,
Tbd, Tsd,  Tl, Tvol, Anz, Thf. Hinweis: BWI3 -> Durchmesser: Tbd, Tsd;
Laenge: Tl. BWI2 -> Durchmesser: Dm; Laenge: Lge, kein Anz.}

\item{ecken}{Dataframe-Tabelle mit Eckenmerkmalen.}

\item{trakte}{Dataframe-tabelle mit Traktmerkmalen.}

\item{bwi}{Angabe zu BWI- Nummer (2 oder 3), um mit denselben Algorithmen
fuer beide Zustaende arbeiten zu koennen.}

\item{krit}{Angabe zu Aufnahmekriterium (2 oder 3), welche verwendet werden
sollen fuer die Berechnung.}

\item{klass}{Liste, mit der die Klassifikationsvariable (Attribut) und ihre
Kategorien definiert werden koennen. Moegliche Klassifikationsattribute sind
Totholzbaumartengruppe <tbagr>, Totholzart <tart>, Zersetzungsgrad <tzg>,
Durchmesser <dm> (Mittendurchmesser, BHD, Stockschnittflaechendurchmessser)
oder Laenge <lge>. z.B. klass = list(attr="tart", kat=list(c(11,12,13),
c(2,3),c(4,5))). Wird furr klass NA uebergeben, erfolgt keine Klassifikation.}

\item{A}{Flaeche in ha des Inventurgebietes zum jeweiligen Inventurzeitpunkt
(sollte konstant sein).}

\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum ueber die Funktion
\code{\link{stratum.fun}} erfolgt. Die Funktion benoetigt des weiteren die
Tabelle <bacode>. Bsp.: list(Wa=c(3,5), Begehbar=1).}

\item{strict}{Logical value which defines wheteher Input Check shall be done
strictly after documentation or not. By default set to FALSE.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und version.totholz.b), \strong{Stratum} (\code{auswahl}),
 \strong{nTE} (Anzahl der Ecken im Stratum), \strong{HBF} (Holzbodenflaeche),
 \strong{se.HBF} (Standardfehler der Holzbodenflaeche), \strong{Klass.Attr}
 (die Klassifizierungsvariable) \strong{Klass.Kat} (Klassifizierungswerte,
 -gruppen), \strong{T.ThVN.klass} (Arrays mit Volumwerten, Gesamtanzahl
 und Fehlern nach den Klassifizierungswerten), \strong{ThVN.ha.klass} (Arrays
 mit Volumenwerten und Gesamsamtanazhl je ha und Fehlern nach den
 Klassifizierungswerten), \strong{nT.klass} (Anzahl der Trakteckenn nach
 Klassifizierungswerten)
}
\description{
Funktion wertet nach  Klassifikation <klass>, Aufnahmekriterium <krit> [2,3]
im Stratum \code{auswahl} aus. Bezueglich der Totholz-Aufnahmeschwellen gibt
es 2 Varianten: \cr
(1) die BWI 2-Kriterien mit Schwellendurchmessern: 20 cm am schwaecheren Ende
bei liegenden Stuecken  bzw. BHD sowie 60 cm Schnittflaechendurchmesser bei
Stoecken \cr
(2) die BWI 3-Kriterien mit Schwellendurchmessern: 10 cm am schwaecheren Ende
bei liegenden Stuecken  bzw. BHD  20 cm Schnittflaechendurchmesser bei
Stoecken.
}
\section{Erstellungsdatum}{
 17.02.2014
}

\section{Aktualisierung}{
 10.04.2014
}

\section{Hinweis}{
 BWI 2-Kriterien bei BWI 3-Aufnahme: Wenn(([Tart]=4 Und
 ([Tbd]>=60 Oder [Tl]>=0,5)) Oder ([Tart]<>4 Und [Tbd]>=20) \cr
 Verallgemeinerte Version fuer die Auswertung unterschiedlicher
 Inventurzeitpunkte, also BWI 2, BWI 3: daher werden die Datentabellen
 \code{totholz}, die Eckenmerkmale \code{ecken} und die \code{trakte} als
 Argumente uebergeben. \cr
 Um Konflikte mit unterschiedlicher Gross-/Kleinschreibung bei den
 Attributnamen zu vermeiden, werden innerhalb dieser Funktion alle
 Attributnamen auf Kleinschreibung umgestellt.
}

\section{TODO}{
 "harmonisieren" -> Basis-Attribute (BWI 2): <TBAGr>, <TArt>,
 <TZg>, <TVol>, <Dm>, <Lge>, <THf>. Neu bei BWI 3: <Tbd> Durchmesser am
 staerkeren Ende bzw. BHD und <Tsd> Durchmesser am duennen Ende (nur bei
 Tart 1 und 13). Fuer Klassifikation relevant ist bei Bruchstuecken der
 Mitten-Durchmesser, bei Stuecken mit Wurzelanlauf (stehend oder liegend) ist
 es der BHD. Mittendurchmesser wird berechnet, wenn Tsd > 0, aus Volumen und
 Laenge. <Dm> = sqrt(tvol/tl/pi)*200; die Bezeichnung <Dm> entspricht auch
 derjenigen der BWI 2. Wegen der Einheitlichkeit mit BWI 2 wird der
 Bezeichner <tl> in <lge> umgewandelt.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
