% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR_Totholz.R
\name{Totholz.Tart.stratum.fun}
\alias{Totholz.Tart.stratum.fun}
\title{Funktion aggregiert Totholz-Volumen nach Totholzart}
\usage{
Totholz.Tart.stratum.fun(totholz, ecken, trakte, bwi, krit, tart.grupp, A,
  auswahl)
}
\arguments{
\item{totholz}{Dataframe-Tabelle mit Totholz-Informationen. Die Tabele muss
mindestens folgende Attribute enthalten: Tnr, Enr, Nr, Tbagr, Tart, Tzg,
Tbd, Tsd,  Tl, Tvol, Anz, Thf. Hinweis: BWI3 -> Durchmesser: Tbd, Tsd;
Laenge: Tl. BWI2 -> Durchmesser: Dm; Laenge: Lge, kein Anz.}

\item{ecken}{Dataframe-Tabelle mit Eckenmerkmalen.}

\item{trakte}{Dataframe-tabelle mit Traktmerkmalen.}

\item{bwi}{Angabe zu BWI- Nummer (2 oder 3), um mit denselben Algorithmen
fuer beide Zustaende arbeiten zu koennen.}

\item{krit}{Angabe zu Aufnahmekriterium (2 oder 3), welche verwendet werden
sollen fuer die Berechnung.}

\item{tart.grupp}{Liste, mit der Totholz-Arten-Gruppen definiert werden
koennen. z.B. tart.grupp = list(g1=c(11,12,13),g2=c(2,3),g3=c(4,5)). Wird
fuer tart.grupp NA uebergeben, werden die Original-tart-Kategorien verwendet,
also bei BWI 3: 11,12,13,2,3,4,5; BWI 2: 1,2,3,4,5.}

\item{A}{Flaeche in ha des Inventurgebietes zum jeweiligen Inventurzeitpunkt
(sollte konstant sein).}

\item{auswahl}{auswahl Liste, welche die Eckenmerkmale mit den Werten
enthaelt, anhand derer die Auswahl fuer das Stratum ueber die Funktion
\code{\link{stratum.fun}} erfolgt. Die Funktion benoetigt des weiteren die
Tabelle <bacode>. Bsp.: list(Wa=c(3,5), Begehbar=1).}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Liste mit
 Erstellungsdatum und version.totholz.b), \strong{Stratum} (\code{auswahl}),
 \strong{nTE} (Anzahl der Ecken im Stratum), \strong{HBF} (Holzbodenflaeche),
 \strong{se.HBF} (Standardfehler der Holzbodenflaeche), \strong{Tart.Liste}
 (Liste mit Angaben zu Totholzarten. Wenn \code{tart.grupp} definiert wurde,
 so wird "Gesamt" ausgegeben. Wenn fuer \code{tart.grupp} NA uebergeben wird,
 so wird eine Liste mit BWI-Codes fuer enthaltende Totholzarten aggregiert),
 \strong{T.ThVN.Tart} (Tabellen mit Volumenwerten, Gesamtanzahl und Fehlern
 nach Totholzart), \strong{ThVN.ha.Tart} (Tabellen mit Volumenwerten,
 Gesamtanzahl je ha und Fehlern nach Totholzart), \strong{nT.Tart}
 (Eckenanzahl nach Totholzart)
}
\description{
Funktion wertet nach  Totholzart <Tart>, Aufnahmekriterium <krit> [2,3] im
Stratum \code{auswahl} aus. Bezueglich der Totholz-Aufnahmeschwellen gibt es
2 Varianten: \cr
(1) die BWI 2-Kriterien mit Schwellendurchmessern: 20 cm am schwaecheren Ende
bei liegenden Stuecken  bzw. BHD sowie 60 cm Schnittflaechendurchmesser bei
Stoecken \cr
(2) die BWI 3-Kriterien mit Schwellendurchmessern: 10 cm am schwaecheren Ende
bei liegenden Stuecken  bzw. BHD  20 cm Schnittflaechendurchmesser bei
Stoecken.
}
\section{Version}{
 1.0 (kae/17.02.2014)
}

\section{Hinweis}{
 BWI 2-Kriterien bei BWI 3-Aufnahme: Wenn(([Tart]=4 Und
 ([Tbd]>=60 Oder [Tl]>=0,5)) Oder ([Tart]<>4 Und [Tbd]>=20) \cr
 Verallgemeinerte Version fuer die Auswertung unterschiedlicher
 Inventurzeitpunkte, also BWI 2, BWI 3: daher werden die Datentabellen
 \code{totholz}, die Eckenmerkmale \code{ecken} und die \code{trakte} als
 Argumente uebergeben. \cr
 Um Konflikte mit unterschiedlicher Gross-/Kleinschreibung bei den
 Attributnamen zu vermeiden, werden innerhalb dieser Funktion alle
 Attributnamen auf Kleinschreibung umgestellt.
}

\section{TODO}{
 "harmonisieren" -> Basis-Attribute (BWI 2): <TBAGr>, <TArt>,
 <TZg>, <TVol>, <Dm>, <Lge>, <THf>. Neu bei BWI 3: <Tbd> Durchmesser am
 staerkeren Ende bzw. BHD und <Tsd> Durchmesser am duennen Ende (nur bei
 Tart 1 und 13). Fuer Klassifikation relevant ist bei Bruchstuecken der
 Mitten-Durchmesser, bei Stuecken mit Wurzelanlauf (stehend oder liegend) ist
 es der BHD. Mittendurchmesser wird berechnet, wenn Tsd > 0, aus Volumen und
 Laenge. <Dm> = sqrt(tvol/tl/pi)*200; die Bezeichnung <Dm> entspricht auch
 derjenigen der BWI 2. Wegen der Einheitlichkeit mit BWI 2 wird der
 Bezeichner <tl> in <lge> umgewandelt.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
