% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R, R/regional.R
\name{tplot}
\alias{tplot}
\alias{tplot}
\title{Change axes, title and legend style of a plot.}
\usage{
tplot(obj)

tplot(obj)
}
\arguments{
\item{obj}{Grafic defined by ggplot().}

\item{obj}{FIXME}
}
\value{
Grafic with predefined axes, title and legend style.
}
\description{
Function changes axes, title and legend styles of a plot. Styles are definded
in \code{\link{theme_bold_axes}}, \code{\link{theme_bold_title}} and
\code{\link{theme_bold_legend}}.

FIXME
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
\keyword{internal}
