% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_species_shares.R
\name{tree_species_percent}
\alias{tree_species_percent}
\title{Creates vector with percentage values for tree species}
\usage{
tree_species_percent(statistics_list)
}
\arguments{
\item{statistics_list}{Name of the list holding the statistical data.}
}
\value{
vektor with tree species areas in percent.
}
\description{
Function selects area for each tree from the statistical list and calculates
the percentages.
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>, Franziska Berg
}
