% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verbiss.bagr.fun}
\alias{verbiss.bagr.fun}
\title{Berechnet mittlere Verbissprozent der Verjuengung}
\usage{
verbiss.bagr.fun(verj, ecken, trakte, auswahl, inv, A,
  bacode = get_data("bacode"))
}
\arguments{
\item{verj}{Tabelle, welche die Kennwerte der Verjuengung mit Schadmerkmal
Verbiss <Biss> enthaelt.}

\item{ecken}{Tabelle mit den Stichproben (Ecken)-Merkmalen fuer die
Startifikation gemaess der Auswahlkriterien in \code{auswahl} (gemaess der
Konvention).}

\item{trakte}{Traktinformationen fuer die Hochrechnung.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{inv}{Inventur: BWI 1: 1; BWI 2, BWI 3: 2.}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}

\item{bacode}{A data.frame containing tree species codes.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Datum),
 \strong{Stratum} (\code{auswahl}), \strong{HBF} (Holzbodenflaeche in ha),
 \strong{se.HBF} (Standardfehler Holzbodenflaeche), \strong{BAGR} (vektor mit
 Baumarten), \strong{m.Verbissproz.BAGR} (Prozentualer Anteil von Verbiss und
 Standardfehler), \strong{BAGR} (Vektor mit Baumarten), \strong{AKL},
 \strong{DKL}, \strong{iVB.bagr.akl.dkl}, \strong{VB.A.bagr.akl.dkl}.
}
\description{
Funktion berechnet das mittlere Verbissprozent der Verjuengung
(Groessenklasse bis 1,3 m) nach BWI-Baumartengruppen entsprechend der
Auswertungskonvention als sogenannter Mean of Ratios (gemaess Absprache mit TI
Petra Henning bzw. Friedrich Schmitz, BMEL vom 11.08.14).
}
\section{Erstellungsdatum}{
 11.08.2014
}

\section{Hinweis}{
 Tabelle \code{bacode} muss geladen sein. \cr
 Folgende Verbiss-Kategorien werden ausgewertet: \cr 0- ohne Verbiss; \cr 1-
 nur Verbiss der Terminalknospe (der letzten 12 Monate); \cr 2- mehrfacher
 Verbiss (auch bei intakter Terminalknospe) ueber einen laengeren Zeitraum
 (einschliesslich der letzten 12 Monate); \cr 12- beide Verbiss-Kategorien
 (1 oder 2).
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
