% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verjg.bagr.fun}
\alias{verjg.bagr.fun}
\title{Aggregiert Stammzahlen der Verjuengung nach Baumartengruppen}
\usage{
verjg.bagr.fun(verj, ecken, trakte, auswahl, inv, A,
  bacode = get_data("bacode"))
}
\arguments{
\item{verj}{Tabelle, welche die Kennwerte der Verjuengung mit Schadmerkmal
Verbiss <Biss> enthaelt.}

\item{ecken}{Tabelle mit den Stichproben (Ecken)-Merkmalen fuer die
Startifikation gemaess der Auswahlkriterien in \code{auswahl} (gemaess der
Konvention).}

\item{trakte}{Traktinformationen fuer die Hochrechnung.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{inv}{Inventur: BWI 1: 1; BWI 2, BWI 3: 2.}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}

\item{bacode}{A data.frame containing tree species codes.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Datum),
 \strong{Stratum} (\code{auswahl}), \strong{HBF} (Holzbodenflaeche in ha),
 \strong{se.HBF} (Standardfehler Holzbodenflaeche), \strong{BAGR} (Vektor mit
 verschiedenen Baumarten), \strong{n.Verjg.BAGR} (2-dimensionalles Array mit
 Informationen zu absoluter Flaeche und Standardfehler (1. Dimension) sowie
 Baumartenanteile in Prozent (2. Dimension) nach Baumartengruppen und fuer
 alle Baumarten).
}
\description{
Funktion aggregiert die Stammzahlen der Verjuengung nach
BWI-Baumartengruppen.
}
\section{Erstellungsdatum}{
 13.08.2014
}

\section{Hinweis}{
 Tabelle \code{bacode} muss geladen sein.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
