% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verjg.bagrupp.fun}
\alias{verjg.bagrupp.fun}
\title{Aggregiert Stammzahlen der Verjuengung nach Baumartengruppen}
\usage{
verjg.bagrupp.fun(verj, ecken, trakte, ba.grupp, auswahl, inv, A)
}
\arguments{
\item{verj}{Tabelle, welche die Kennwerte der Verjuengung mit Schadmerkmal
Verbiss <Biss> enthaelt.}

\item{ecken}{Tabelle mit den Stichproben (Ecken)-Merkmalen fuer die
Startifikation gemaess der Auswahlkriterien in \code{auswahl} (gemaess der
Konvention).}

\item{trakte}{Traktinformationen fuer die Hochrechnung.}

\item{ba.grupp}{Liste mit Baumarten-Zusammenfassungen zu Baumgruppen mit
Bezeichner der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa",
"DglKiLae", "Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51), c(100), c(110,111), c(112:199), c(200:299))).}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{inv}{Inventur: BWI 1: 1; BWI 2, BWI 3: 2.}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Datum),
 \strong{Stratum} (\code{auswahl}), \strong{HBF} (Holzbodenflaeche in ha),
 \strong{se.HBF} (Standardfehler Holzbodenflaeche), \strong{BAGR} (Labels
 fuer Baumartengruppen aus \code{ba.grupp}), \strong{n.Verjg.BAGR}
 (2-dimensionalles Array mit Informationen zu absoluter
 Flaeche und Standardfehler (1. Dimension) sowie Baumartenanteile in Prozent
 (2. Dimension) nach Baumartengruppen und fuer alle Baumarten).
}
\description{
Funktion aggregiert die Stammzahlen der Verjuengung nach frei definierbaren
Baumartengruppen.
}
\section{Erstellungsdatum}{
 13.08.2014
}

\section{Hinweis}{
 Tabelle \code{bacode} muss geladen sein.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
