% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verjg.kl4.bagr.fun}
\alias{verjg.kl4.bagr.fun}
\title{Berechnet Verjuengungsflaechen nach Baumartengruppen und Verjuengungsart}
\usage{
verjg.kl4.bagr.fun(verj.kl4, ecken, trakte, auswahl, A,
  bacode = get_data("bacode"))
}
\arguments{
\item{verj.kl4}{Tabelle mit den Attributen der Bestockungsansprache <= 4m.}

\item{ecken}{Stichprobenmerkmale.}

\item{trakte}{Traktparameter.}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}

\item{bacode}{A data.frame containing tree species codes.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Datum),
 \strong{Stratum} (\code{auswahl}), \strong{HBF} (Holzbodenflaeche in ha),
 \strong{se.HBF} (Standardfehler Holzbodenflaeche), \strong{BAGR} (Vektor mit
 verschiedenen Baumarten), \strong{VjgArt} (Vektor mit verschiedenen
 Verjuengungsarten), \strong{Verjg.kl4m.BAF,VArt.BAGR} (absolute Flaeche vom
 Inventurgebiet und Standardfehler nach Verjuengungsart und ueber alle
 Baumarten, alle Verjuengungsarten zusammen und nach Baumartengruppen,
 alle Verjuengungsrten und alle Baumartengruppen zusammen).
}
\description{
Funktion wertet die Verjuengungsaufnahme im PK (Probebaumkennziffer)
Bestockung <= 4m Hoehe aus. Es werden die Flaechen aus Bedeckungsgrad- und
Baumartenanteilsschaetzung differenziert nach BWI-Baumartengruppen und
Verjuengungsart berechnet.
}
\section{Erstellungsdatum}{
 13.08.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
