% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verjg.kl4.bagrupp.fun}
\alias{verjg.kl4.bagrupp.fun}
\title{Berechnet Verjuengungsflaechen nach Baumartengruppen und Verjuengungsart}
\usage{
verjg.kl4.bagrupp.fun(verj.kl4, ecken, trakte, ba.grupp, auswahl, A)
}
\arguments{
\item{verj.kl4}{Tabelle mit den Attributen der Bestockungsansprache <= 4m.}

\item{ecken}{Stichprobenmerkmale.}

\item{trakte}{Traktparameter.}

\item{ba.grupp}{Liste mit BA-Zusammenfassungen zu Baumgruppen mit Bezeichner
der Baumarten-Gruppen ("lab") z.B. list(bagr.lab = c("FiTa", "DglKiLae",
"Bu", "Ei", "BLb", "WLb"), ba.grupp =list(c(10:19,30:39,90:99),
c(20:29,40,50,51),c(100),c(110,111), c(112:199), c(200:299))).}

\item{auswahl}{Liste, welche die Eckenmerkmale mit den Werten enthaelt,
anhand derer die Auswahl fuer das Stratum erfolgt. Bsp.: list(Wa=c(3,5),
Begehbar=1).}

\item{A}{Gesamtflaeche des Inventurgebietes in ha.}
}
\value{
Liste mit folgenden Komponenten: \strong{Log} (Datum),
 \strong{Stratum} (\code{auswahl}), \strong{HBF} (Holzbodenflaeche in ha),
 \strong{se.HBF} (Standardfehler Holzbodenflaeche), \strong{BAGR} (Labels
 fuer Baumartengruppen aus \code{ba.grupp}), \strong{VjgArt} (Vektor mit verschiedenen
 Verjuengungsarten), \strong{Verjg.kl4m.BAF,VArt.BAGR} (absolute Flaeche vom
 Inventurgebiet und Standardfehler nach Verjuengungsart und ueber alle
 Baumarten, alle Verjuengungsarten zusammen und nach Baumartengruppen,
 alle Verjuengungsrten und alle Baumartengruppen zusammen).
}
\description{
Funktion wertet die Verjuengungsaufnahme im PK (Probebaumkennziffer)
Bestockung <= 4m Hoehe aus. Es werden die Flaechen aus Bedeckungsgrad- und
Baumartenanteilsschaetzung differenziert nach frei definierbaren
BWI-Baumartengruppen und Verjuengungsart berechnet.
}
\section{Erstellungsdatum}{
 14.08.2014
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
