% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWI3_HR.R
\name{verjg.tab.fun}
\alias{verjg.tab.fun}
\title{Erstelle Dataframe-Tabelle fuer Verjuengung}
\usage{
verjg.tab.fun(vj.tab)
}
\arguments{
\item{vj.tab}{Ergebnistabelle aus \code{\link{verjg.bagr.fun}} oder
\code{\link{verjg.bagrupp.fun}}.}
}
\value{
Dataframe-Tabelle.
}
\description{
Funktion erzeugt aus Ergebnistabelle der Funktion \code{\link{verjg.bagr.fun}}
bzw. \code{\link{verjg.bagrupp.fun}} eine Dataframe-Tabelle.
}
\section{Note}{
 Hilfsfunktion zur Tabellenausgabe.
}

\author{
Gerald Kaendler \email{gerald.kaendler@forst.bwl.de}
}
