% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{xtable_by_group}
\alias{xtable_by_group}
\title{Xtable by group}
\usage{
xtable_by_group(data, subs, group = "Baumartengruppe",
  label_prefix = "total")
}
\arguments{
\item{data}{Output table of \code{\link{all_data_by_group}}}

\item{subs}{String which defines subset parameters, e.g. "Eigentumsart ==
'Privatwald'".}

\item{group}{String which defines the column in data which shall be used for the
grouping on the x-axis. by default "Baumartengruppe"}

\item{label_prefix}{String which defines the prefix for teh label.}
}
\value{
Fix me.
}
\description{
Fix me
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
