% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundeswaldinventur-package.R
\docType{package}
\name{bundeswaldinventur-package}
\alias{bundeswaldinventur-package}
\title{Analysing the Bundeswaldinventur in the State of Baden-Wuerttemberg}
\description{
A collection of design-based estimators and reporting functions for
substrata of the area of interest.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_bundeswaldinventur", package =
"bundeswaldinventur")}.
}
\keyword{package}
