% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundeswaldinventur-package.R
\docType{package}
\name{bundeswaldinventur-package}
\alias{bundeswaldinventur-package}
\title{What it Does (One Line, Title Case)}
\description{
A description is a paragraph consisting of one or more sentences.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_bundeswaldinventur", package =
"bundeswaldinventur")}.
}
\keyword{package}
