% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Load Data from \pkg{bwibw}}
\usage{
get_data(name, package = get_options("data_source"))
}
\arguments{
\item{name}{a string of length one giving the name of the data.frame.}

\item{package}{The package to load data from.}
}
\value{
the data requested.
}
\description{
Get data if we want to use example data provided with the this of the
\pkg{bwibw} package.
}
\details{
Allows to globally en/dis-abling the reading of example
data into function arguments, Which we need because Gerald used global
data.frames in his functions.
}
\examples{
if ("bwibw" \%in\% rownames(installed.packages())) {
   get_data("bacode", package = "bwibw") 
} else {
   get_data("bacode") 
}
}
