% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_getters.R
\name{get_design}
\alias{get_design}
\title{get attributes from the underlying sampling design}
\usage{
get_design(attribute, bwi, hard_coded = TRUE)
}
\arguments{
\item{attribute}{The attribute of interest.}

\item{bwi}{The BWI to use.}

\item{hard_coded}{use hardcoded values or try to retrieve them from the
respective trakte data.frame?}
}
\value{
an integer
}
\description{
get attributes from the underlying sampling design
}
\examples{
get_design("nte", "3")
get_design("nTE", "bwi3")
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
