% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_deadwood}
\alias{plot_deadwood}
\title{FIXME}
\usage{
plot_deadwood(deadwood_2, deadwood_3_2, deadwood_3, deadwood_2a, deadwood_3_2a,
  deadwood_3a, graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii)
}
\arguments{
\item{deadwood_2}{FIXME}

\item{deadwood_3_2}{FIXME}

\item{deadwood_3}{FIXME}

\item{deadwood_2a}{FIXME}

\item{deadwood_3_2a}{FIXME}

\item{deadwood_3a}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
