% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_group_area_by_age}
\alias{plot_group_area_by_age}
\title{FIXME}
\usage{
plot_group_area_by_age(b1, b2, b3, a, do_errors_relative = FALSE,
  graphic_width = get_options("graphics_width"),
  graphic_height = graphics_height, graphic_directory = graphics_directory,
  file_name_district = regional_file_name,
  title_district = krs.grupp$string[i], have_title = FALSE)
}
\arguments{
\item{b1}{FIXME}

\item{b2}{FIXME}

\item{b3}{FIXME}

\item{a}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{title_district}{FIXME}

\item{have_title}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
