% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_groups_stocks}
\alias{plot_groups_stocks}
\title{FIXME}
\usage{
plot_groups_stocks(b1, b2, b3, do_errors_relative = FALSE,
  graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii, species_groups_labels, have_title = FALSE)
}
\arguments{
\item{b1}{FIXME}

\item{b2}{FIXME}

\item{b3}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}

\item{species_groups_labels}{FIXME}

\item{have_title}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
