% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_growth_loss}
\alias{plot_growth_loss}
\title{FIXME}
\usage{
plot_growth_loss(g_l_12, g_l_23, g_l_12_all, do_errors_relative = FALSE,
  graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii, species_groups_labels)
}
\arguments{
\item{g_l_12}{FIXME}

\item{g_l_23}{FIXME}

\item{g_l_12_all}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}

\item{species_groups_labels}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
