% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_loss}
\alias{plot_loss}
\title{FIXME}
\usage{
plot_loss(loss1, loss2, loss1_all, loss2_all,
  graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii, have_title = FALSE)
}
\arguments{
\item{loss1}{FIXME}

\item{loss2}{FIXME}

\item{loss1_all}{FIXME}

\item{loss2_all}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}

\item{have_title}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
