% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_ntns}
\alias{plot_ntns}
\title{FIXME}
\usage{
plot_ntns(ntns2, ntns3, graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii, have_title = FALSE)
}
\arguments{
\item{ntns2}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}

\item{have_title}{FIXME}

\item{ntns1}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
