% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{plot_silviculturally_relevant}
\alias{plot_silviculturally_relevant}
\title{FIXME}
\usage{
plot_silviculturally_relevant(relevant_species_2, relevant_species_3,
  do_errors_relative = FALSE, graphic_width = get_options("graphics_width"),
  graphic_height = get_options("graphics_height"), graphic_directory,
  file_name_district, dot_district_tex, plots_directory, title_district,
  stratum_index, stratii, species_groups_labels, have_title = FALSE)
}
\arguments{
\item{relevant_species_2}{FIXME}

\item{relevant_species_3}{FIXME}

\item{do_errors_relative}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{plots_directory}{FIXME}

\item{title_district}{FIXME}

\item{stratum_index}{FIXME}

\item{stratii}{FIXME}

\item{species_groups_labels}{FIXME}

\item{have_title}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
