% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_and_tables.R
\name{reassemble_by_group}
\alias{reassemble_by_group}
\title{Reassembles by ownership and a variable abbreviation}
\usage{
reassemble_by_group(owner, abbreviation, data_prefix = "FVBN.bagr")
}
\arguments{
\item{owner}{String which defines the owner.}

\item{abbreviation}{String which defines the variable which shall be
predicted.}

\item{data_prefix}{The prefix from which the lists names are to be
established.}
}
\value{
Dataframe.
}
\description{
Gather a variable for all three BWI from statistics list that are compiled
for and named after a special ownership.
}
\examples{
assign('fvbn.bagr.gw.3', envir = .GlobalEnv, 
FVBN.bagrupp.akl.dkl.stratum.fun.2a(
        get_data("baeume.3"), get_data("ecken.3"), 
                  get_data("trakte.3"), get_design("a", 3), 2, get_bwi_species_groups(),
 			       list(A.ob = 500, A.b = 500), list(D.unt = 0, D.ob = 500,
 			       D.b = 500, Ndh = FALSE),
 			       list(Wa = c(3, 5), Begehbar = 1))
)
assign('fvbn.bagr.gw.2', envir = .GlobalEnv, 
FVBN.bagrupp.akl.dkl.stratum.fun.2a(
        get_data("baeume.2"), get_data("ecken.2"), 
                  get_data("trakte.2"), get_design("a", 2), 2, get_bwi_species_groups(),

 			       list(A.ob = 500, A.b = 500), list(D.unt = 0, D.ob = 500,
 			       D.b = 500, Ndh = FALSE),
 			       list(Wa = c(1, 2, 3), Begehbar = 1))
)
assign('fvbn.bagr.gw.1', envir = .GlobalEnv, 
FVBN.bagrupp.akl.dkl.stratum.fun.2a(
        get_data("baeume.1"), get_data("ecken.1"), 
                  get_data("trakte.1"), get_design("a", 1), 1, get_bwi_species_groups(),
 			       list(A.ob = 500, A.b = 500), list(D.unt = 0, D.ob = 500,
 			       D.b = 500, Ndh = FALSE),
 			       list(Wa = c(1, 2, 3), Begehbar = 1))
)
reassemble_by_group("gw", "BAF", data_prefix = "fvbn.bagr")
}
\author{
Dominik Cullmann <dominik.cullmann@forst.bwl.de>
}
