% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{regeneration}
\alias{regeneration}
\title{FIXME}
\usage{
regeneration(r2_a, r2_bs1, r2_bs2, r3_a, r3_bs1, r3_bs2,
  graphic_width = get_options("graphics_width"),
  graphic_height = graphics_height, graphic_directory = graphics_directory,
  file_name_district = regional_file_name,
  title_district = krs.grupp$string[i])
}
\arguments{
\item{r2_a}{FIXME}

\item{r2_bs1}{FIXME}

\item{r2_bs2}{FIXME}

\item{r3_a}{FIXME}

\item{r3_bs1}{FIXME}

\item{r3_bs2}{FIXME}

\item{graphic_width}{FIXME}

\item{graphic_height}{FIXME}

\item{graphic_directory}{FIXME}

\item{file_name_district}{FIXME}

\item{title_district}{FIXME}
}
\description{
FIXME
}
\keyword{internal}
